sui_core::accumulators

Function build_accumulator_barrier_tx

pub fn build_accumulator_barrier_tx(
    epoch: u64,
    accumulator_root_obj_initial_shared_version: SequenceNumber,
    checkpoint_height: u64,
    settlement_effects: &[TransactionEffects],
) -> TransactionKind

Builds the barrier transaction that advances the version of the accumulator root object. This must be called after all settlement transactions have been executed. settlement_effects contains the effects of all settlement transactions.