sui_core

Module authority

Modules

authority_per_epoch_store

authority_per_epoch_store_pruner

authority_store_pruner

authority_store_tables

authority_store_types

authority_test_utils

backpressure

consensus_tx_status_cache

epoch_start_configuration

execution_time_estimator

shared_object_congestion_tracker

shared_object_version_manager

submitted_transaction_cache

test_authority_builder

transaction_deferral

transaction_reject_reason_cache

Structs

AuthorityMetrics

Prometheus metrics which can be displayed in Grafana, queried and alerted on

AuthorityState

AuthorityStore

ALL_OBJ_VER determines whether we want to store all past versions of every object in the store. Authority doesn’t store them, but other entities such as replicas will. S is a template on Authority signature state. This allows SuiDataStore to be used on either authorities or non-authorities. Specifically, when storing transactions and effects, S allows SuiDataStore to either store the authority signed version or unsigned version.

ExecutionEnv

Execution env contains the “environment” for the transaction to be executed in, that is, all the information necessary for execution that is not specified by the transaction itself.

ForkRecoveryState

NodeStateDump

ObjDumpFormat

RandomnessRoundReceiver

ResolverWrapper

A wrapper to make Orphan Rule happy

Enums

UpdateType

Constants

DEV_INSPECT_GAS_COIN_VALUE

WAIT_FOR_FASTPATH_INPUT_TIMEOUT

Type Aliases

StableSyncAuthoritySigner

a Trait object for Signer that is: