sui_core::authority::authority_per_epoch_store
Struct AuthorityEpochTablesReadOnly

pub struct AuthorityEpochTablesReadOnly {

    pub signed_transactions: DBMap<TransactionDigest, TrustedEnvelope<SenderSignedData, AuthoritySignInfo>>,
    pub owned_object_locked_transactions: DBMap<ObjectRef, LockDetailsWrapper>,
    pub effects_signatures: DBMap<TransactionDigest, AuthoritySignInfo>,
    pub signed_effects_digests: DBMap<TransactionDigest, TransactionEffectsDigest>,
    pub transaction_cert_signatures: DBMap<TransactionDigest, AuthorityStrongQuorumSignInfo>,
    pub next_shared_object_versions_v2: DBMap<ConsensusObjectSequenceKey, SequenceNumber>,
    pub consensus_message_processed: DBMap<SequencedConsensusTransactionKey, bool>,
    pub pending_consensus_transactions: DBMap<ConsensusTransactionKey, ConsensusTransaction>,
    pub last_consensus_stats: DBMap<u64, ExecutionIndicesWithStats>,
    pub reconfig_state: DBMap<u64, ReconfigState>,
    pub end_of_publish: DBMap<AuthorityName, ()>,
    pub builder_digest_to_checkpoint: DBMap<TransactionDigest, CheckpointSequenceNumber>,
    pub transaction_key_to_digest: DBMap<TransactionKey, TransactionDigest>,
    pub pending_checkpoint_signatures: DBMap<(CheckpointSequenceNumber, u64), CheckpointSignatureMessage>,
    pub builder_checkpoint_summary_v2: DBMap<CheckpointSequenceNumber, BuilderCheckpointSummary>,
    pub state_hash_by_checkpoint: DBMap<CheckpointSequenceNumber, GlobalStateHash>,
    pub running_root_state_hash: DBMap<CheckpointSequenceNumber, GlobalStateHash>,
    pub authority_capabilities: DBMap<AuthorityName, AuthorityCapabilitiesV1>,
    pub authority_capabilities_v2: DBMap<AuthorityName, AuthorityCapabilitiesV2>,
    pub override_protocol_upgrade_buffer_stake: DBMap<u64, u64>,
    pub executed_transactions_to_checkpoint: DBMap<TransactionDigest, CheckpointSequenceNumber>,
    pub pending_jwks: DBMap<(AuthorityName, JwkId, JWK), ()>,
    pub active_jwks: DBMap<(u64, (JwkId, JWK)), ()>,
    pub deferred_transactions: DBMap<DeferralKey, Vec<VerifiedSequencedConsensusTransaction>>,
    pub dkg_processed_messages_v2: DBMap<PartyId, VersionedProcessedMessage>,
    pub dkg_used_messages_v2: DBMap<u64, VersionedUsedProcessedMessages>,
    pub dkg_confirmations_v2: DBMap<PartyId, VersionedDkgConfirmation>,
    pub dkg_output: DBMap<u64, Output<G2Element, G2Element>>,
    pub randomness_next_round: DBMap<u64, RandomnessRound>,
    pub randomness_highest_completed_round: DBMap<u64, RandomnessRound>,
    pub randomness_last_round_timestamp: DBMap<u64, TimestampMs>,
    pub congestion_control_object_debts: DBMap<ObjectID, CongestionPerObjectDebt>,
    pub congestion_control_randomness_object_debts: DBMap<ObjectID, CongestionPerObjectDebt>,
    pub execution_time_observations: DBMap<(u64, AuthorityIndex), Vec<(ExecutionTimeObservationKey, Duration)>>,

}

Fields
signed_transactions: DBMap<TransactionDigest, TrustedEnvelope<SenderSignedData, AuthoritySignInfo>>owned_object_locked_transactions: DBMap<ObjectRef, LockDetailsWrapper>effects_signatures: DBMap<TransactionDigest, AuthoritySignInfo>signed_effects_digests: DBMap<TransactionDigest, TransactionEffectsDigest>transaction_cert_signatures: DBMap<TransactionDigest, AuthorityStrongQuorumSignInfo>next_shared_object_versions_v2: DBMap<ConsensusObjectSequenceKey, SequenceNumber>consensus_message_processed: DBMap<SequencedConsensusTransactionKey, bool>pending_consensus_transactions: DBMap<ConsensusTransactionKey, ConsensusTransaction>last_consensus_stats: DBMap<u64, ExecutionIndicesWithStats>reconfig_state: DBMap<u64, ReconfigState>end_of_publish: DBMap<AuthorityName, ()>builder_digest_to_checkpoint: DBMap<TransactionDigest, CheckpointSequenceNumber>transaction_key_to_digest: DBMap<TransactionKey, TransactionDigest>pending_checkpoint_signatures: DBMap<(CheckpointSequenceNumber, u64), CheckpointSignatureMessage>builder_checkpoint_summary_v2: DBMap<CheckpointSequenceNumber, BuilderCheckpointSummary>state_hash_by_checkpoint: DBMap<CheckpointSequenceNumber, GlobalStateHash>running_root_state_hash: DBMap<CheckpointSequenceNumber, GlobalStateHash>authority_capabilities: DBMap<AuthorityName, AuthorityCapabilitiesV1>authority_capabilities_v2: DBMap<AuthorityName, AuthorityCapabilitiesV2>override_protocol_upgrade_buffer_stake: DBMap<u64, u64>executed_transactions_to_checkpoint: DBMap<TransactionDigest, CheckpointSequenceNumber>pending_jwks: DBMap<(AuthorityName, JwkId, JWK), ()>active_jwks: DBMap<(u64, (JwkId, JWK)), ()>deferred_transactions: DBMap<DeferralKey, Vec<VerifiedSequencedConsensusTransaction>>dkg_processed_messages_v2: DBMap<PartyId, VersionedProcessedMessage>dkg_used_messages_v2: DBMap<u64, VersionedUsedProcessedMessages>dkg_confirmations_v2: DBMap<PartyId, VersionedDkgConfirmation>dkg_output: DBMap<u64, Output<G2Element, G2Element>>randomness_next_round: DBMap<u64, RandomnessRound>randomness_highest_completed_round: DBMap<u64, RandomnessRound>randomness_last_round_timestamp: DBMap<u64, TimestampMs>congestion_control_object_debts: DBMap<ObjectID, CongestionPerObjectDebt>congestion_control_randomness_object_debts: DBMap<ObjectID, CongestionPerObjectDebt>execution_time_observations: DBMap<(u64, AuthorityIndex), Vec<(ExecutionTimeObservationKey, Duration)>>
Implementations

impl AuthorityEpochTablesReadOnly

pub fn open_tables_read_only( primary_path: PathBuf, with_secondary_path: Option<PathBuf>, metric_conf: MetricConf, global_db_options_override: Option<Options>, ) -> Self

Open in read only mode. No limitation on number of processes to do this

pub fn dump( &self, cf_name: &str, page_size: u16, page_number: usize, ) -> Result<BTreeMap<String, String>>

Dump all key-value pairs in the page at the given table name Tables must be opened in read only mode using open_tables_read_only

pub fn table_summary(&self, table_name: &str) -> Result<TableSummary>

Get key value sizes from the db Tables must be opened in read only mode using open_tables_read_only

pub fn describe_tables() -> BTreeMap<String, (String, String)>

pub fn try_catch_up_with_primary_all(&self) -> Result<()>

Try catch up with primary for all tables. This can be a slow operation Tables must be opened in read only mode using open_tables_read_only