Struct sui_core::authority::authority_per_epoch_store::AuthorityPerEpochStore

pub struct AuthorityPerEpochStore { /* private fields */ }

Implementations

impl AuthorityPerEpochStore

pub fn new(
name: AuthorityName,
committee: Committee,
parent_path: &Path,
db_options: Option<Options>,
metrics: Arc<EpochMetrics>
) -> Arc<Self>

pub fn get_parent_path(&self) -> PathBuf

pub fn new_at_next_epoch(
&self,
name: AuthorityName,
new_committee: Committee
) -> Arc<Self>

pub fn wal(
&self
) -> &Arc<DBWriteAheadLog<TrustedCertificate, (InnerTemporaryStore, TrustedSignedTransactionEffects)>>

pub fn committee(&self) -> &Committee

pub fn epoch(&self) -> EpochId

pub async fn acquire_tx_guard(
&self,
cert: &VerifiedCertificate
) -> SuiResult<DBTxGuard<'_, TrustedCertificate, (InnerTemporaryStore, TrustedSignedTransactionEffects)>>

pub async fn acquire_tx_lock(&self, digest: &TransactionDigest) -> CertLockGuard

Acquire the lock for a tx without writing to the WAL.

pub fn store_reconfig_state(&self, new_state: &ReconfigState) -> SuiResult

pub fn insert_transaction(
&self,
transaction: VerifiedSignedTransaction
) -> SuiResult

pub fn get_signed_transaction(
&self,
tx_digest: &TransactionDigest
) -> SuiResult<Option<VerifiedSignedTransaction>>

pub fn multi_get_next_shared_object_versions<'a>(
&self,
ids: impl Iterator<Item = &'a ObjectID>
) -> SuiResult<Vec<Option<SequenceNumber>>>

pub fn get_last_checkpoint_boundary(&self) -> (u64, Option<u64>)

pub fn get_last_consensus_index(&self) -> SuiResult<ExecutionIndicesWithHash>

pub fn get_transactions_in_checkpoint_range(
&self,
from_height_excluded: Option<u64>,
to_height_included: u64
) -> SuiResult<Vec<TransactionDigest>>

pub fn get_pending_certificate(
&self,
tx: &TransactionDigest
) -> Result<Option<VerifiedCertificate>, TypedStoreError>

pending_certificates table related methods. Should only be used from TransactionManager. Gets one certificate pending execution.

pub fn all_pending_certificates(&self) -> SuiResult<Vec<VerifiedCertificate>>

Gets all pending certificates. Used during recovery.

pub fn remove_pending_certificate(
&self,
digest: &TransactionDigest
) -> SuiResult<()>

Deletes one pending certificate.

pub fn get_all_pending_consensus_transactions(
&self
) -> Vec<ConsensusTransaction>

pub fn get_shared_locks(
&self,
transaction_digest: &TransactionDigest
) -> Result<Vec<(ObjectID, SequenceNumber)>, SuiError>

Read shared object locks / versions for a specific transaction.

pub async fn set_assigned_shared_object_versions(
&self,
certificate: &VerifiedCertificate,
assigned_versions: &Vec<(ObjectID, SequenceNumber)>,
parent_sync_store: impl ParentSync
) -> SuiResult

pub async fn acquire_shared_locks_from_effects(
&self,
certificate: &VerifiedCertificate,
effects: &TransactionEffects,
parent_sync_store: impl ParentSync
) -> SuiResult

Lock a sequence number for the shared objects of the input transaction based on the effects of that transaction. Used by full nodes, which don’t listen to consensus.

pub fn insert_checkpoint_boundary(&self, index: u64, height: u64) -> SuiResult

pub fn insert_pending_consensus_transactions(
&self,
transaction: &ConsensusTransaction,
lock: Option<&RwLockReadGuard<'_, ReconfigState>>
) -> SuiResult

When submitting a certificate caller must provide a ReconfigState lock guard and verify that it allows new user certificates

pub fn remove_pending_consensus_transaction(
&self,
key: &ConsensusTransactionKey
) -> SuiResult

pub fn pending_consensus_certificates_empty(&self) -> bool

pub fn pending_consensus_certificates(&self) -> HashSet<TransactionDigest>

pub fn insert_pending_certificates(
&self,
certs: &[VerifiedCertificate]
) -> Result<(), TypedStoreError>

Stores a list of pending certificates to be executed.

pub fn is_tx_cert_consensus_message_processed(
&self,
certificate: &CertifiedTransaction
) -> SuiResult<bool>

Check whether certificate was processed by consensus. For shared lock certificates, if this function returns true means shared locks for this certificate are set

pub fn is_consensus_message_processed(
&self,
key: &ConsensusTransactionKey
) -> SuiResult<bool>

pub async fn consensus_message_processed_notify(
&self,
key: ConsensusTransactionKey
) -> Result<(), SuiError>

pub fn has_sent_end_of_publish(
&self,
authority: &AuthorityName
) -> SuiResult<bool>

pub async fn user_signatures_for_checkpoint(
&self,
digests: &[TransactionDigest]
) -> SuiResult<Vec<Signature>>

Note: this is async function as it waits for certificates to be processed by consensus before returning

pub fn record_end_of_publish(
&self,
authority: AuthorityName,
key: ConsensusTransactionKey,
consensus_index: ExecutionIndicesWithHash
) -> SuiResult

Returns Ok(true) if 2f+1 end of publish messages were recorded at this point

pub async fn record_owned_object_cert_from_consensus(
&self,
transaction: &ConsensusTransaction,
certificate: &VerifiedCertificate,
consensus_index: ExecutionIndicesWithHash
) -> Result<(), SuiError>

Caller is responsible to call consensus_message_processed before this method

pub async fn record_shared_object_cert_from_consensus(
&self,
transaction: &ConsensusTransaction,
certificate: &VerifiedCertificate,
consensus_index: ExecutionIndicesWithHash,
parent_sync_store: impl ParentSync
) -> Result<(), SuiError>

Locks a sequence number for the shared objects of the input transaction. Also updates the last consensus index, consensus_message_processed and pending_certificates tables. This function must only be called from the consensus task (i.e. from handle_consensus_transaction).

Caller is responsible to call consensus_message_processed before this method

pub fn record_consensus_transaction_processed(
&self,
transaction: &ConsensusTransaction,
consensus_index: ExecutionIndicesWithHash
) -> Result<(), SuiError>

pub fn record_certified_checkpoint_signatures(
&self,
certified_checkpoint: &CertifiedCheckpointSummary
) -> Result<(), TypedStoreError>

Called when a ceritified checkpoint is inserted into the checkpoint store. This function records the validators’ participation in the certified checkpoint by incrementing the counters stored in the epoch tables. Tallying score metrics are also updated at this point to send the most up-to-date scores.

pub fn finish_consensus_certificate_process(
&self,
key: ConsensusTransactionKey,
certificate: &VerifiedCertificate,
consensus_index: ExecutionIndicesWithHash
) -> SuiResult

pub fn finish_assign_shared_object_versions(
&self,
key: ConsensusTransactionKey,
certificate: &VerifiedCertificate,
consensus_index: ExecutionIndicesWithHash,
assigned_versions: Vec<(ObjectID, SequenceNumber)>,
next_versions: Vec<(ObjectID, SequenceNumber)>
) -> SuiResult

pub fn test_insert_user_signature(
&self,
digest: TransactionDigest,
signature: &Signature
)

pub fn final_epoch_checkpoint(&self) -> SuiResult<Option<u64>>

pub fn last_checkpoint(
&self,
to_height_included: u64
) -> SuiResult<Option<(u64, Vec<TransactionDigest>)>>

Returns transaction digests from consensus_message_order table in the “checkpoint range”.

Checkpoint range is defined from the last seen checkpoint(excluded) to the provided to_height (included)

pub fn record_checkpoint_boundary(&self, commit_round: u64) -> SuiResult

pub fn get_reconfig_state_read_lock_guard(
&self
) -> RwLockReadGuard<'_, ReconfigState>

pub fn get_reconfig_state_write_lock_guard(
&self
) -> RwLockWriteGuard<'_, ReconfigState>

pub fn close_user_certs(&self, lock_guard: RwLockWriteGuard<'_, ReconfigState>)

pub async fn epoch_terminated(&self)

Notify epoch is terminated, can only be called once on epoch store

pub async fn wait_epoch_terminated(&self)

Waits for the notification about epoch termination

pub async fn within_alive_epoch<F: Future + Send>(
&self,
f: F
) -> Result<F::Output, ()>

This function executes given future until epoch_terminated is called If future finishes before epoch_terminated is called, future result is returned If epoch_terminated is called before future is resolved, error is returned

In addition to the early termination guarantee, this function also prevents epoch_terminated() if future is being executed.

pub fn handle_commit_boundary<C: CheckpointServiceNotify>(
&self,
committed_dag: &Arc<CommittedSubDag>,
checkpoint_service: &Arc<C>
) -> SuiResult

pub fn get_pending_checkpoints(
&self
) -> Vec<(CheckpointCommitHeight, PendingCheckpoint)>

pub fn get_pending_checkpoint(
&self,
index: &CheckpointCommitHeight
) -> Result<Option<PendingCheckpoint>, TypedStoreError>

pub fn insert_pending_checkpoint(
&self,
index: &CheckpointCommitHeight,
checkpoint: &PendingCheckpoint
) -> Result<(), TypedStoreError>

pub fn process_pending_checkpoint(
&self,
commit_height: CheckpointCommitHeight,
content_info: &[(CheckpointSummary, CheckpointContents)]
) -> Result<(), TypedStoreError>

pub fn put_genesis_checkpoint_in_builder(
&self,
summary: &CheckpointSummary,
contents: &CheckpointContents
) -> SuiResult<()>

Register genesis checkpoint in builder DB

pub fn last_built_checkpoint_summary(
&self
) -> SuiResult<Option<(CheckpointSequenceNumber, CheckpointSummary)>>

pub fn get_built_checkpoint_summary(
&self,
sequence: CheckpointSequenceNumber
) -> SuiResult<Option<CheckpointSummary>>

pub fn builder_included_transaction_in_checkpoint(
&self,
digest: &TransactionDigest
) -> Result<bool, TypedStoreError>

pub fn get_pending_checkpoint_signatures_iter(
&self,
checkpoint_seq: CheckpointSequenceNumber,
starting_index: u64
) -> Result<impl Iterator<Item = ((CheckpointSequenceNumber, u64), CheckpointSignatureMessage)> + '_, TypedStoreError>

pub fn get_last_checkpoint_signature_index(&self) -> u64

pub fn get_num_certified_checkpoint_sigs_by(
&self,
name: &AuthorityName
) -> SuiResult<Option<(Option<CheckpointSequenceNumber>, u64)>>

pub fn insert_checkpoint_signature(
&self,
checkpoint_seq: CheckpointSequenceNumber,
index: u64,
info: &CheckpointSignatureMessage
) -> Result<(), TypedStoreError>

pub fn record_epoch_reconfig_start_time_metric(&self)