sui_core::authority

Module authority_store_types

Structs

StoreObjectValue

Forked version of sui_types::object::Object Used for efficient storing of move objects in the database

Enums

StoreData

Forked version of sui_types::object::Data Adds extra enum value IndirectObject, which represents a reference to an object stored separately

StoreObjectV1

StoreObjectWrapper

Enum wrapper for versioning

Functions

get_store_object

Type Aliases

StoreObject