sui_core::authority::epoch_start_configuration

Trait EpochStartConfigTrait

pub trait EpochStartConfigTrait {
    // Required methods
    fn epoch_digest(&self) -> CheckpointDigest;
    fn epoch_start_state(&self) -> &EpochStartSystemState;
    fn flags(&self) -> &[EpochFlag];
    fn bridge_committee_initiated(&self) -> bool;
    fn system_object_initial_shared_version(
        &self,
        object_id: ObjectID,
    ) -> Option<SequenceNumber>;
}

Required Methods

fn epoch_digest(&self) -> CheckpointDigest

fn epoch_start_state(&self) -> &EpochStartSystemState

fn flags(&self) -> &[EpochFlag]

fn bridge_committee_initiated(&self) -> bool

fn system_object_initial_shared_version( &self, object_id: ObjectID, ) -> Option<SequenceNumber>

Returns the initial shared version of the given system object as of the start of the epoch, or None if the object did not exist yet.