Struct sui_core::authority::AuthorityState

pub struct AuthorityState {
    pub name: AuthorityName,
    pub committee: Committee,
    pub secret: StableSyncAuthoritySigner,
    pub consensus_guardrail: AtomicUsize,
    /* private fields */
}

Fields
name: AuthorityName

The name of this authority.

committee: Committee

Committee of this Sui instance.

secret: StableSyncAuthoritySigner

The signature key of the authority.

consensus_guardrail: AtomicUsize

Ensures there can only be a single consensus client is updating the state.

Implementations

impl AuthorityState

The authority state encapsulates all state, drives execution, and ensures safety.

Note the authority operations can be accessed through a read ref (&) and do not require &mut. Internally a database is synchronized through a mutex lock.

Repeating valid commands should produce no changes and return no error.

pub fn subscribe_batch(&self) -> BroadcastReceiver

Get a broadcast receiver for updates

pub async fn handle_transaction(
&self,
transaction: Transaction
) -> Result<TransactionInfoResponse, SuiError>

Initiate a new transaction.

pub async fn handle_confirmation_transaction(
&self,
confirmation_transaction: ConfirmationTransaction
) -> SuiResult<TransactionInfoResponse>

Confirm a transfer.

pub async fn handle_consensus_certificate(
&self,
certificate: CertifiedTransaction,
last_consensus_index: ExecutionIndices
) -> SuiResult<()>

Process certificates coming from the consensus. It is crucial that this function is only called by a single task (ie. the task handling consensus outputs).

pub async fn handle_transaction_info_request(
&self,
request: TransactionInfoRequest
) -> Result<TransactionInfoResponse, SuiError>

pub async fn handle_account_info_request(
&self,
request: AccountInfoRequest
) -> Result<AccountInfoResponse, SuiError>

pub async fn handle_object_info_request(
&self,
request: ObjectInfoRequest
) -> Result<ObjectInfoResponse, SuiError>

pub async fn handle_batch_info_request(
&self,
request: BatchInfoRequest
) -> Result<(VecDeque<UpdateItem>, (bool, TxSequenceNumber, TxSequenceNumber)), SuiError>

Handles a request for a batch info. It returns a sequence of [batches, transactions, batches, transactions] as UpdateItems, and a flag that if true indicates the request goes beyond the last batch in the database.

pub async fn new(
committee: Committee,
name: AuthorityName,
secret: StableSyncAuthoritySigner,
store: Arc<AuthorityStore>,
genesis_packages: Vec<Vec<CompiledModule>>,
genesis_ctx: &mut TxContext
) -> Self

pub async fn new_without_genesis(
committee: Committee,
name: AuthorityName,
secret: StableSyncAuthoritySigner,
store: Arc<AuthorityStore>
) -> Self

pub async fn insert_genesis_object(&self, object: Object)

pub async fn insert_genesis_objects_bulk_unsafe(&self, objects: &[&Object])

pub async fn set_transaction_lock(
&self,
mutable_input_objects: &[ObjectRef],
signed_transaction: SignedTransaction
) -> Result<(), SuiError>

Set the transaction lock to a specific transaction

pub async fn get_transaction_lock(
&self,
object_ref: &ObjectRef
) -> Result<Option<SignedTransaction>, SuiError>

Get a read reference to an object/seq lock

pub async fn read_certificate(
&self,
digest: &TransactionDigest
) -> Result<Option<CertifiedTransaction>, SuiError>

Read from the DB of certificates

pub async fn parent(&self, object_ref: &ObjectRef) -> Option<TransactionDigest>

pub async fn get_objects(
&self,
_objects: &[ObjectID]
) -> Result<Vec<Option<Object>>, SuiError>

pub async fn get_parent_iterator(
&self,
object_id: ObjectID,
seq: Option<SequenceNumber>
) -> Result<impl Iterator<Item = (ObjectRef, TransactionDigest)> + '_, SuiError>

Returns all parents (object_ref and transaction digests) that match an object_id, at any object version, or optionally at a specific version.

pub async fn get_latest_parent_entry(
&self,
object_id: ObjectID
) -> Result<Option<(ObjectRef, TransactionDigest)>, SuiError>

impl AuthorityState

pub fn last_batch(&self) -> Result<Option<SignedBatch>, SuiError>

pub fn init_batches_from_database(&mut self) -> Result<AuthorityBatch, SuiError>

Initializes the database to handle batches, and recovers from a potential crash by creating a last batch to include any trailing trasnactions not in a batch.

This needs exclusive access to the database at this point, so we take the authority state as a &mut.

pub async fn run_batch_service(
&self,
min_batch_size: u64,
max_delay: Duration
) -> SuiResult<()>

pub async fn handle_batch_streaming(
&self,
request: BatchInfoRequest
) -> Result<impl Stream<Item = Result<BatchInfoResponseItem, SuiError>>, SuiError>