Struct sui_core::authority::AuthorityStore

pub struct AuthorityStore { /* private fields */ }

ALL_OBJ_VER determines whether we want to store all past versions of every object in the store. Authority doesn’t store them, but other entities such as replicas will. S is a template on Authority signature state. This allows SuiDataStore to be used on either authorities or non-authorities. Specifically, when storing transactions and effects, S allows SuiDataStore to either store the authority signed version or unsigned version.

Implementations

impl AuthorityStore

pub async fn open( perpetual_tables: Arc<AuthorityPerpetualTables>, genesis: &Genesis, committee_store: &Arc<CommitteeStore>, indirect_objects_threshold: usize, enable_epoch_sui_conservation_check: bool, registry: &Registry ) -> SuiResult<Arc<Self>>

Open an authority store by directory path. If the store is empty, initialize it using genesis.

pub fn update_epoch_flags_metrics(&self, old: &[EpochFlag], new: &[EpochFlag])

pub async fn open_with_committee_for_testing( perpetual_tables: Arc<AuthorityPerpetualTables>, committee: &Committee, genesis: &Genesis, indirect_objects_threshold: usize ) -> SuiResult<Arc<Self>>

pub fn get_root_state_hash( &self, epoch: EpochId ) -> SuiResult<ECMHLiveObjectSetDigest>

pub fn get_root_state_accumulator( &self, epoch: EpochId ) -> (CheckpointSequenceNumber, Accumulator)

pub fn get_recovery_epoch_at_restart(&self) -> SuiResult<EpochId>

pub fn get_effects( &self, effects_digest: &TransactionEffectsDigest ) -> SuiResult<Option<TransactionEffects>>

pub fn effects_exists( &self, effects_digest: &TransactionEffectsDigest ) -> SuiResult<bool>

Returns true if we have an effects structure for this transaction digest

pub fn multi_get_events( &self, event_digests: &[TransactionEventsDigest] ) -> SuiResult<Vec<Option<TransactionEvents>>>

pub fn multi_get_effects<'a>( &self, effects_digests: impl Iterator<Item = &'a TransactionEffectsDigest> ) -> SuiResult<Vec<Option<TransactionEffects>>>

pub fn get_executed_effects( &self, tx_digest: &TransactionDigest ) -> SuiResult<Option<TransactionEffects>>

pub fn multi_get_executed_effects_digests( &self, digests: &[TransactionDigest] ) -> SuiResult<Vec<Option<TransactionEffectsDigest>>>

Given a list of transaction digests, returns a list of the corresponding effects only if they have been executed. For transactions that have not been executed, None is returned.

pub fn multi_get_executed_effects( &self, digests: &[TransactionDigest] ) -> SuiResult<Vec<Option<TransactionEffects>>>

Given a list of transaction digests, returns a list of the corresponding effects only if they have been executed. For transactions that have not been executed, None is returned.

pub fn is_tx_already_executed( &self, digest: &TransactionDigest ) -> SuiResult<bool>

pub fn get_deleted_shared_object_previous_tx_digest( &self, object_id: &ObjectID, version: &SequenceNumber, epoch_id: EpochId ) -> Result<Option<TransactionDigest>, TypedStoreError>

pub fn get_last_shared_object_deletion_info( &self, object_id: &ObjectID, epoch_id: EpochId ) -> SuiResult<Option<(SequenceNumber, TransactionDigest)>>

pub async fn notify_read_root_state_hash( &self, epoch: EpochId ) -> SuiResult<(CheckpointSequenceNumber, Accumulator)>

Returns future containing the state hash for the given epoch once available

pub fn deprecated_insert_finalized_transactions( &self, digests: &[TransactionDigest], epoch: EpochId, sequence: CheckpointSequenceNumber ) -> SuiResult

pub fn deprecated_is_transaction_executed_in_checkpoint( &self, digest: &TransactionDigest ) -> SuiResult<bool>

pub fn deprecated_get_transaction_checkpoint( &self, digest: &TransactionDigest ) -> SuiResult<Option<(EpochId, CheckpointSequenceNumber)>>

pub fn deprecated_multi_get_transaction_checkpoint( &self, digests: &[TransactionDigest] ) -> SuiResult<Vec<Option<(EpochId, CheckpointSequenceNumber)>>>

pub fn database_is_empty(&self) -> SuiResult<bool>

Returns true if there are no objects in the database

pub fn get_object_ref_prior_to_key( &self, object_id: &ObjectID, version: VersionNumber ) -> Result<Option<ObjectRef>, SuiError>

pub fn multi_get_object_by_key( &self, object_keys: &[ObjectKey] ) -> Result<Vec<Option<Object>>, SuiError>

pub fn multi_get_object_with_more_accurate_error_return( &self, object_refs: &[ObjectRef] ) -> Result<Vec<Object>, SuiError>

Load a list of objects from the store by object reference. If they exist in the store, they are returned directly. If any object missing, we try to figure out the best error to return. If the object we are asking is currently locked at a future version, we know this transaction is out-of-date and we return a ObjectVersionUnavailableForConsumption, which indicates this is not retriable. Otherwise, we return a ObjectNotFound error, which indicates this is retriable.

pub fn get_objects( &self, objects: &[ObjectID] ) -> Result<Vec<Option<Object>>, SuiError>

Get many objects

pub fn have_received_object_at_version( &self, object_id: &ObjectID, version: VersionNumber, epoch_id: EpochId ) -> Result<bool, SuiError>

pub fn have_deleted_owned_object_at_version_or_after( &self, object_id: &ObjectID, version: VersionNumber, epoch_id: EpochId ) -> Result<bool, SuiError>

pub fn get_input_object_keys( &self, digest: &TransactionDigest, objects: &[InputObjectKind], epoch_store: &AuthorityPerEpochStore ) -> BTreeSet<InputKey>

Gets the input object keys from input object kinds, by determining the versions of owned, shared and package objects. When making changes, please see if check_sequenced_input_objects() below needs similar changes as well.

pub fn multi_input_objects_available( &self, keys: impl Iterator<Item = InputKey> + Clone, receiving_objects: HashSet<InputKey>, epoch_store: &AuthorityPerEpochStore ) -> Result<Vec<bool>, SuiError>

Checks if the input object identified by the InputKey exists, with support for non-system packages i.e. when version is None. If the input object doesn’t exist and it’s a receiving object, we also check if the object exists in the object marker table and view it as existing if it is in the table.

pub async fn execution_lock_for_executable_transaction( &self, transaction: &VerifiedExecutableTransaction ) -> SuiResult<RwLockReadGuard<'_, EpochId>>

Attempts to acquire execution lock for an executable transaction. Returns the lock if the transaction is matching current executed epoch Returns None otherwise

pub async fn execution_lock_for_reconfiguration( &self ) -> RwLockWriteGuard<'_, EpochId>

pub async fn insert_objects_unsafe_for_testing_only( &self, objects: &[Object] ) -> SuiResult

NOTE: this function is only to be used for fuzzing and testing. Never use in prod

pub fn bulk_insert_live_objects( perpetual_db: &AuthorityPerpetualTables, live_objects: impl Iterator<Item = LiveObject>, indirect_objects_threshold: usize, expected_sha3_digest: &[u8; 32] ) -> SuiResult<()>

pub async fn set_epoch_start_configuration( &self, epoch_start_configuration: &EpochStartConfiguration ) -> SuiResult

pub fn get_epoch_start_configuration( &self ) -> SuiResult<Option<EpochStartConfiguration>>

pub async fn update_state( &self, inner_temporary_store: InnerTemporaryStore, transaction: &VerifiedTransaction, effects: &TransactionEffects, epoch_id: EpochId ) -> SuiResult

Updates the state resulting from the execution of a certificate.

Internally it checks that all locks for active inputs are at the correct version, and then writes objects, certificates, parents and clean up locks atomically.

pub fn check_owned_object_locks_exist(&self, objects: &[ObjectRef]) -> SuiResult

Checks multiple object locks exist. Returns UserInputError::ObjectNotFound if cannot find lock record for at least one of the objects. Returns UserInputError::ObjectVersionUnavailableForConsumption if at least one object lock is not initialized at the given version.

pub fn initialize_locks( locks_table: &DBMap<ObjectRef, Option<LockDetailsWrapper>>, write_batch: &mut DBBatch, objects: &[ObjectRef], is_force_reset: bool ) -> SuiResult

pub async fn revert_state_update( &self, tx_digest: &TransactionDigest ) -> SuiResult

This function is called at the end of epoch for each transaction that’s executed locally on the validator but didn’t make to the last checkpoint. The effects of the execution is reverted here. The following things are reverted:

All new object states are deleted.

owner_index table change is reverted.

NOTE: transaction and effects are intentionally not deleted. It’s possible that if this node is behind, the network will execute the transaction in a later epoch. In that case, we need to keep it saved so that when we receive the checkpoint that includes it from state sync, we are able to execute the checkpoint. TODO: implement GC for transactions that are no longer needed.

pub fn find_object_lt_or_eq_version( &self, object_id: ObjectID, version: SequenceNumber ) -> Option<Object>

Return the object with version less then or eq to the provided seq number. This is used by indexer to find the correct version of dynamic field child object. We do not store the version of the child object, but because of lamport timestamp, we know the child must have version number less then or eq to the parent.

pub fn get_latest_object_ref_or_tombstone( &self, object_id: ObjectID ) -> Result<Option<ObjectRef>, SuiError>

Returns the latest object reference we have for this object_id in the objects table.

The method may also return the reference to a deleted object with a digest of ObjectDigest::deleted() or ObjectDigest::wrapped() and lamport version of a transaction that deleted the object. Note that a deleted object may re-appear if the deletion was the result of the object being wrapped in another object.

If no entry for the object_id is found, return None.

pub fn get_latest_object_ref_if_alive( &self, object_id: ObjectID ) -> Result<Option<ObjectRef>, SuiError>

Returns the latest object reference if and only if the object is still live (i.e. it does not return tombstones)

pub fn get_latest_object_or_tombstone( &self, object_id: ObjectID ) -> Result<Option<(ObjectKey, StoreObjectWrapper)>, SuiError>

Returns the latest object we have for this object_id in the objects table.

If no entry for the object_id is found, return None.

pub fn insert_transaction_and_effects( &self, transaction: &VerifiedTransaction, transaction_effects: &TransactionEffects ) -> Result<(), TypedStoreError>

pub fn multi_insert_transaction_and_effects<'a>( &self, transactions: impl Iterator<Item = &'a VerifiedExecutionData> ) -> Result<(), TypedStoreError>

pub fn multi_get_transaction_blocks( &self, tx_digests: &[TransactionDigest] ) -> SuiResult<Vec<Option<VerifiedTransaction>>>

pub fn get_transaction_block( &self, tx_digest: &TransactionDigest ) -> Result<Option<VerifiedTransaction>, TypedStoreError>

pub fn get_transactions_and_serialized_sizes<'a>( &self, digests: impl IntoIterator<Item = &'a TransactionDigest> ) -> Result<Vec<Option<(VerifiedTransaction, usize)>>, TypedStoreError>

pub fn get_sui_system_state_object(&self) -> SuiResult<SuiSystemState>

pub fn iter_live_object_set( &self, include_wrapped_object: bool ) -> impl Iterator<Item = LiveObject> + '_

pub fn expensive_check_sui_conservation( self: &Arc<Self>, old_epoch_store: &AuthorityPerEpochStore ) -> SuiResult

pub fn expensive_check_is_consistent_state( &self, checkpoint_executor: &CheckpointExecutor, accumulator: Arc<StateAccumulator>, cur_epoch_store: &AuthorityPerEpochStore, panic: bool )