Struct sui_core::authority::AuthorityTemporaryStore

pub struct AuthorityTemporaryStore<S> { /* private fields */ }

Implementations

impl<S> AuthorityTemporaryStore<S>

pub fn new(
package_store: Arc<S>,
input_objects: Vec<(InputObjectKind, Object)>,
tx_digest: TransactionDigest
) -> Self

Creates a new store associated with an authority store, and populates it with initial objects.

pub fn objects(&self) -> &BTreeMap<ObjectID, Object>

pub fn written(&self) -> &BTreeMap<ObjectID, (ObjectRef, Object)>

pub fn deleted(&self) -> &BTreeMap<ObjectID, (SequenceNumber, DeleteKind)>

pub fn into_inner(
self
) -> (BTreeMap<ObjectID, Object>, Vec<ObjectRef>, BTreeMap<ObjectID, (ObjectRef, Object)>, BTreeMap<ObjectID, (SequenceNumber, DeleteKind)>, Vec<Event>)

Break up the structure and return its internal stores (objects, active_inputs, written, deleted)

pub fn ensure_active_inputs_mutated(&mut self)

For every object from active_inputs (i.e. all mutable objects), if they are not mutated during the transaction execution, force mutating them by incrementing the sequence number. This is required to achieve safety. We skip the last object, which is always the gas object, because gas object will be updated after this.

pub fn charge_gas_for_storage_changes(
&mut self,
gas_status: &mut SuiGasStatus<'_>,
gas_object: &mut Object
) -> SuiResult

For every object changes, charge gas accordingly. Since by this point we haven’t charged gas yet, the gas object hasn’t been mutated yet. Passing in gas_object_size so that we can also charge for the gas object mutation in advance.

pub fn to_effects(
&self,
shared_object_refs: Vec<ObjectRef>,
transaction_digest: &TransactionDigest,
transaction_dependencies: Vec<TransactionDigest>,
status: ExecutionStatus,
gas_object_id: &ObjectID
) -> TransactionEffects