Struct sui_core::authority::SuiDataStore

pub struct SuiDataStore<const ALL_OBJ_VER: bool, S> {
    pub executed_sequence: DBMap<TxSequenceNumber, TransactionDigest>,
    pub batches: DBMap<TxSequenceNumber, SignedBatch>,
    /* private fields */
}

ALL_OBJ_VER determines whether we want to store all past versions of every object in the store. Authority doesn’t store them, but other entities such as replicas will. S is a template on Authority signature state. This allows SuiDataStore to be used on either authorities or non-authorities. Specifically, when storing transactions and effects, S allows SuiDataStore to either store the authority signed version or unsigned version.

Fields
executed_sequence: DBMap<TxSequenceNumber, TransactionDigest>

A sequence on all executed certificates and effects.

batches: DBMap<TxSequenceNumber, SignedBatch>

A sequence of batches indexing into the sequence of executed transactions.

Implementations

impl<const ALL_OBJ_VER: bool, S: Eq + Serialize + for<'de> Deserialize<'de>> SuiDataStore<ALL_OBJ_VER, S>

pub fn open<P: AsRef<Path>>(path: P, db_options: Option<Options>) -> Self

Open an authority store by directory path

pub fn get_effects(
&self,
transaction_digest: &TransactionDigest
) -> SuiResult<TransactionEffects>

Returns the TransactionEffects if we have an effects structure for this transaction digest

pub fn effects_exists(
&self,
transaction_digest: &TransactionDigest
) -> SuiResult<bool>

Returns true if we have an effects structure for this transaction digest

pub fn transaction_exists(
&self,
transaction_digest: &TransactionDigest
) -> SuiResult<bool>

Returns true if we have a transaction structure for this transaction digest

pub fn database_is_empty(&self) -> SuiResult<bool>

Returns true if there are no objects in the database

pub fn next_sequence_number(&self) -> Result<TxSequenceNumber, SuiError>

pub fn get_account_objects(
&self,
account: SuiAddress
) -> Result<Vec<ObjectRef>, SuiError>

pub fn get_object(
&self,
object_id: &ObjectID
) -> Result<Option<Object>, SuiError>

Read an object and return it, or Err(ObjectNotFound) if the object was not found.

pub fn get_objects(
&self,
_objects: &[ObjectID]
) -> Result<Vec<Option<Object>>, SuiError>

Get many objects

pub fn get_transaction_lock(
&self,
object_ref: &ObjectRef
) -> Result<Option<TransactionEnvelope<S>>, SuiError>

Read a lock or returns Err(TransactionLockDoesNotExist) if the lock does not exist.

pub fn read_certificate(
&self,
digest: &TransactionDigest
) -> Result<Option<CertifiedTransaction>, SuiError>

Read a certificate and return an option with None if it does not exist.

pub fn parent(
&self,
object_ref: &ObjectRef
) -> Result<Option<TransactionDigest>, SuiError>

Read the transactionDigest that is the parent of an object reference (ie. the transaction that created an object at this version.)

pub fn multi_get_parents(
&self,
object_refs: &[ObjectRef]
) -> Result<Vec<Option<TransactionDigest>>, SuiError>

Batch version of parent function.

pub fn get_parent_iterator(
&self,
object_id: ObjectID,
seq: Option<SequenceNumber>
) -> Result<impl Iterator<Item = (ObjectRef, TransactionDigest)> + '_, SuiError>

Returns all parents (object_ref and transaction digests) that match an object_id, at any object version, or optionally at a specific version.

pub fn sequenced<'a>(
&self,
transaction_digest: &TransactionDigest,
object_ids: impl Iterator<Item = &'a ObjectID>
) -> Result<Vec<Option<SequenceNumber>>, SuiError>

Read a lock for a specific (transaction, shared object) pair.

pub fn all_shared_locks(
&self,
transaction_digest: &TransactionDigest
) -> Result<Vec<(ObjectID, SequenceNumber)>, SuiError>

Read a lock for a specific (transaction, shared object) pair.

pub fn insert_genesis_object(&self, object: Object) -> SuiResult

Insert a genesis object.

pub fn insert_object_direct(
&self,
object_ref: ObjectRef,
object: &Object
) -> SuiResult

Insert an object directly into the store, and also update relevant tables, including initiating the transaction lock. This is used by the gateway to insert object directly. TODO: We need this today because we don’t have another way to sync an account.

pub fn bulk_object_insert(&self, objects: &[&Object]) -> SuiResult<()>

This function is used by the bench.rs script, and should not be used in other contexts In particular it does not check the old locks before inserting new ones, so the objects must be new.

pub fn set_transaction_lock(
&self,
owned_input_objects: &[ObjectRef],
transaction: TransactionEnvelope<S>
) -> Result<(), SuiError>

Set the transaction lock to a specific transaction

This function checks all locks exist, are either None or equal to the passed transaction and then sets them to the transaction. Otherwise an Err is returned. Locks are set atomically in this implementation.

pub fn update_state<BackingPackageStore>(
&self,
temporary_store: AuthorityTemporaryStore<BackingPackageStore>,
certificate: &CertifiedTransaction,
effects: &TransactionEffectsEnvelope<S>,
sequence_number: Option<TxSequenceNumber>
) -> SuiResult

Updates the state resulting from the execution of a certificate.

Internally it checks that all locks for active inputs are at the correct version, and then writes locks, objects, certificates, parents atomically.

pub fn update_objects_state_for_genesis<BackingPackageStore>(
&self,
temporary_store: AuthorityTemporaryStore<BackingPackageStore>,
transaction_digest: TransactionDigest
) -> Result<(), SuiError>

Persist temporary storage to DB for genesis modules

pub fn update_gateway_state(
&self,
active_inputs: Vec<(InputObjectKind, Object)>,
mutated_objects: HashMap<ObjectRef, Object>,
certificate: CertifiedTransaction,
effects: TransactionEffectsEnvelope<S>,
sequence_number: GatewayTxSeqNumber
) -> SuiResult

This is used by the Gateway to update its local store after a transaction succeeded on the authorities. Since we don’t yet have local execution on the gateway, we will need to recreate the temporary store based on the inputs and effects to update it properly.

pub fn get_latest_parent_entry(
&self,
object_id: ObjectID
) -> Result<Option<(ObjectRef, TransactionDigest)>, SuiError>

Returns the last entry we have for this object in the parents_sync index used to facilitate client and authority sync. In turn the latest entry provides the latest object_reference, and also the latest transaction that has interacted with this object.

This parent_sync index also contains entries for deleted objects (with a digest of ObjectDigest::deleted()), and provides the transaction digest of the certificate that deleted the object. Note that a deleted object may re-appear if the deletion was the result of the object being wrapped in another object.

If no entry for the object_id is found, return None.

pub fn remove_shared_objects_locks(
&self,
write_batch: DBBatch,
transaction_digest: &TransactionDigest,
transaction: &CertifiedTransaction
) -> SuiResult<DBBatch>

Remove the shared objects locks. This function is not safety-critical and is only need to cleanup the store.

pub fn persist_certificate_and_lock_shared_objects(
&self,
certificate: CertifiedTransaction,
consensus_index: ExecutionIndices
) -> Result<(), SuiError>

Lock a sequence number for the shared objects of the input transaction. Also update the last consensus index.

pub fn transactions_in_seq_range(
&self,
start: GatewayTxSeqNumber,
end: GatewayTxSeqNumber
) -> SuiResult<Vec<(GatewayTxSeqNumber, TransactionDigest)>>

pub fn batches_and_transactions(
&self,
start: u64,
end: u64
) -> Result<(Vec<SignedBatch>, Vec<(TxSequenceNumber, TransactionDigest)>), SuiError>

Retrieves batches including transactions within a range.

This function returns all signed batches that enclose the requested transaction including the batch preceding the first requested transaction, the batch including the last requested transaction (if there is one) and all batches in between.

Transactions returned include all transactions within the batch that include the first requested transaction, all the way to at least all the transactions that are included in the last batch returned. If the last requested transaction is outside a batch (one has not yet been generated) the function returns all transactions at the end of the sequence that are in TxSequenceOrder (and ignores any that are out of order.)

pub fn last_consensus_index(&self) -> SuiResult<ExecutionIndices>

Return the latest consensus index. It is used to bootstrap the consensus client.

pub fn get_transaction(
&self,
transaction_digest: &TransactionDigest
) -> SuiResult<Option<TransactionEnvelope<S>>>

pub fn get_certified_transaction(
&self,
transaction_digest: &TransactionDigest
) -> SuiResult<Option<CertifiedTransaction>>

impl<const A: bool> SuiDataStore<A, AuthoritySignInfo>

pub fn get_signed_transaction_info(
&self,
transaction_digest: &TransactionDigest
) -> Result<TransactionInfoResponse, SuiError>

impl<const A: bool> SuiDataStore<A, EmptySignInfo>

pub fn pending_transactions(&self) -> &DBMap<TransactionDigest, Transaction>