Struct sui_core::authority::TemporaryStore

pub struct TemporaryStore<S> { /* private fields */ }

Implementations

impl<S> TemporaryStore<S>

pub fn new( store: S, input_objects: InputObjects, tx_digest: TransactionDigest, protocol_config: &ProtocolConfig ) -> TemporaryStore<S>

Creates a new store associated with an authority store, and populates it with initial objects.

pub fn new_for_mock_transaction( store: S, input_objects: InputObjects, tx_digest: TransactionDigest, protocol_config: &ProtocolConfig ) -> TemporaryStore<S>

WARNING! Should only be used for dry run and dev inspect! In dry run and dev inspect, you might load a dynamic field that is actually too new for the transaction. Ideally, we would want to load the “correct” dynamic fields, but as that is not easily determined, we instead set the lamport version MAX, which is a valid lamport version for any object used in the transaction (preventing internal assertions or invariant violations from being triggered)

pub fn objects(&self) -> &BTreeMap<ObjectID, Object, Global>

pub fn into_inner(self) -> InnerTemporaryStore

Break up the structure and return its internal stores (objects, active_inputs, written, deleted)

pub fn to_effects( self, shared_object_refs: Vec<(ObjectID, SequenceNumber, ObjectDigest), Global>, transaction_digest: &TransactionDigest, transaction_dependencies: Vec<TransactionDigest, Global>, gas_cost_summary: GasCostSummary, status: ExecutionStatus, gas: &[(ObjectID, SequenceNumber, ObjectDigest)], epoch: u64 ) -> (InnerTemporaryStore, TransactionEffects)

pub fn write_object(&mut self, object: Object, kind: WriteKind)

pub fn smash_gas( &mut self, gas: &[(ObjectID, SequenceNumber, ObjectDigest)] ) -> Result<(ObjectID, SequenceNumber, ObjectDigest), ExecutionError>

pub fn delete_object( &mut self, id: &ObjectID, version: SequenceNumber, kind: DeleteKind )

pub fn drop_writes(&mut self)

pub fn reset( &mut self, gas: &[(ObjectID, SequenceNumber, ObjectDigest)], gas_status: &mut SuiGasStatus )

Resets any mutations, deletions, and events recorded in the store, as well as any storage costs and rebates, then Re-runs gas smashing. Effects on store are now as if we were about to begin execution

pub fn log_event(&mut self, event: Event)

pub fn read_object(&self, id: &ObjectID) -> Option<&Object>

pub fn apply_object_changes( &mut self, changes: BTreeMap<ObjectID, ObjectChange, Global> )

pub fn save_loaded_child_objects( &mut self, loaded_child_objects: BTreeMap<ObjectID, SequenceNumber, Global> )

pub fn estimate_effects_size_upperbound(&self) -> usize

pub fn written_objects_size(&self) -> usize

pub fn conserve_unmetered_storage_rebate( &mut self, unmetered_storage_rebate: u64 )

If there are unmetered storage rebate (due to system transaction), we put them into the storage rebate of 0x5 object.

impl<S> TemporaryStore<S>where S: ObjectStore,

pub fn check_ownership_invariants( &self, sender: &SuiAddress, gas: &[(ObjectID, SequenceNumber, ObjectDigest)], is_epoch_change: bool ) -> Result<(), SuiError>

impl<S> TemporaryStore<S>where S: ObjectStore,

pub fn charge_gas_legacy<T>( &mut self, gas_object_id: ObjectID, gas_status: &mut SuiGasStatus, execution_result: &mut Result<T, ExecutionError>, gas: &[(ObjectID, SequenceNumber, ObjectDigest)] )

Compute tx storage gas costs and tx storage rebates, update storage_rebate field of mutated objects

Deduct computation gas costs and storage costs to gas_object_id, credit storage rebates to gas_object_id. gas_object_id can be None if this is a system transaction.

impl<S> TemporaryStore<S>where S: ObjectStore,

pub fn charge_gas<T>( &mut self, gas_object_id: Option<ObjectID>, gas_status: &mut SuiGasStatus, execution_result: &mut Result<T, ExecutionError>, gas: &[(ObjectID, SequenceNumber, ObjectDigest)] ) -> GasCostSummary

Compute tx storage gas costs and tx storage rebates, update storage_rebate field of mutated objects

Deduct computation gas costs and storage costs to gas_object_id, credit storage rebates to gas_object_id.

impl<S> TemporaryStore<S>where S: ObjectStore,

pub fn advance_epoch_safe_mode( &mut self, params: &AdvanceEpochParams, protocol_config: &ProtocolConfig )

impl<S> TemporaryStore<S>where S: GetModule + ObjectStore + BackingPackageStore,

pub fn check_sui_conserved( &self, advance_epoch_gas_summary: Option<(u64, u64)>, layout_resolver: &mut impl LayoutResolver, do_expensive_checks: bool ) -> Result<(), ExecutionError>

Check that this transaction neither creates nor destroys SUI. This should hold for all txes except the epoch change tx, which mints staking rewards equal to the gas fees burned in the previous epoch. Specifically, this checks two key invariants about storage fees and storage rebate:

all SUI in storage rebate fields of input objects should flow either to the transaction storage rebate, or the transaction non-refundable storage rebate

all SUI charged for storage should flow into the storage rebate field of some output object

If do_expensive_checks is true, this will also check a third invariant:

all SUI in input objects (including coins etc in the Move part of an object) should flow either to an output object, or be burned as part of computation fees or non-refundable storage rebate

This function is intended to be called after we have charged for gas + applied the storage rebate to the gas object, but before we have updated object versions. If do_expensive_checks is false, this function will only check conservation of object storage rea epoch_fees and epoch_rebates are only set for advance epoch transactions. The advance epoch transaction would mint epoch_fees amount of SUI, and burn epoch_rebates amount of SUI. We need these information for conservation check.