Struct sui_core::authority_aggregator::AuthorityAggregator

pub struct AuthorityAggregator<A> {
    pub committee: Committee,
    pub authority_clients: BTreeMap<AuthorityName, SafeClient<A>>,
}

Fields
committee: Committee

Our Sui committee.

authority_clients: BTreeMap<AuthorityName, SafeClient<A>>

How to talk to this committee.

Implementations

impl<A> AuthorityAggregator<A>

pub fn new(
committee: Committee,
authority_clients: BTreeMap<AuthorityName, A>
) -> Self

impl<A> AuthorityAggregator<A> where
A: AuthorityAPI + Send + Sync + 'static + Clone,

pub async fn sync_authority_source_to_destination(
&self,
cert: ConfirmationTransaction,
source_authority: AuthorityName,
destination_authority: AuthorityName
) -> Result<(), SuiError>

Sync a certificate and all its dependencies to a destination authority, using a source authority to get information about parent certificates.

Note: Both source and destination may be byzantine, therefore one should always time limit the call to this function to avoid byzantine authorities consuming an unbounded amount of resources.

pub async fn sync_all_given_objects(
&self,
objects: &[ObjectID],
timeout_after_quorum: Duration
) -> Result<(Vec<(Object, Option<MoveStructLayout>, Option<CertifiedTransaction>)>, Vec<(ObjectRef, Option<CertifiedTransaction>)>), SuiError>

Takes a list of object IDs, goes to all (quorum+timeout) of authorities to find their latest version, and then updates all authorities with the latest version of each object.

pub async fn sync_all_owned_objects(
&self,
address: SuiAddress,
timeout_after_quorum: Duration
) -> Result<(Vec<(Object, Option<MoveStructLayout>, Option<CertifiedTransaction>)>, Vec<(ObjectRef, Option<CertifiedTransaction>)>), SuiError>

Ask authorities for the user owned objects. Then download all objects at all versions present on authorities, along with the certificates preceding them, and update lagging authorities to the latest version of the object.

This function returns all objects, including those that are no more owned by the user (but were previously owned by the user), as well as a list of deleted object references.

pub async fn process_transaction(
&self,
transaction: Transaction,
timeout_after_quorum: Duration
) -> Result<CertifiedTransaction, SuiError>

Takes a transaction, brings all authorities up to date with the versions of the objects needed, and then submits the transaction to make a certificate.

pub async fn execute_transaction(
&self,
transaction: &Transaction
) -> Result<(CertifiedTransaction, TransactionEffects), Error>

pub async fn get_object_info_execute(
&self,
object_id: ObjectID
) -> SuiResult<ObjectRead>

pub fn fetch_objects_from_authorities(
&self,
object_refs: BTreeSet<ObjectRef>
) -> Receiver<SuiResult<Object>>

Given a list of object refs, download the objects.