Trait sui_core::authority_aggregator::TransactionCertifier

pub trait TransactionCertifier: Sync + Send + 'static {
    fn create_certificate<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        transaction: &'life1 VerifiedTransaction,
        self_state: &'life2 AuthorityState,
        timeout: Duration
    ) -> Pin<Box<dyn Future<Output = Result<VerifiedCertificate>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;
}

This trait provides a method for an authority to get a certificate from the network for a specific transaction. In order to create a certificate, we need to create the network authority aggregator based on the Sui system state committee/network information. This is needed to create a certificate for the advance epoch transaction during reconfiguration. However to make testing easier, we sometimes want to use local authority clients that do not involve full-fledged network Sui nodes (e.g. when we want to abstract out Narwhal). In order to support both network clients and local clients, this trait is defined to hide the difference. We implement this trait for both NetworkTransactionCertifier and LocalTransactionCertifier.

Required Methods

fn create_certificate<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
transaction: &'life1 VerifiedTransaction,
self_state: &'life2 AuthorityState,
timeout: Duration
) -> Pin<Box<dyn Future<Output = Result<VerifiedCertificate>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

This function first loads the Sui system state object from self_state, get the committee information, creates an AuthorityAggregator, and use the aggregator to create a certificate for the specified transaction.