Trait sui_core::authority_client::AuthorityAPI

pub trait AuthorityAPI {
    fn handle_transaction<'life0, 'async_trait>(
        &'life0 self, 
        transaction: Transaction
    ) -> Pin<Box<dyn Future<Output = Result<TransactionInfoResponse, SuiError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn handle_confirmation_transaction<'life0, 'async_trait>(
        &'life0 self, 
        transaction: ConfirmationTransaction
    ) -> Pin<Box<dyn Future<Output = Result<TransactionInfoResponse, SuiError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn handle_consensus_transaction<'life0, 'async_trait>(
        &'life0 self, 
        transaction: ConsensusTransaction
    ) -> Pin<Box<dyn Future<Output = Result<TransactionInfoResponse, SuiError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn handle_account_info_request<'life0, 'async_trait>(
        &'life0 self, 
        request: AccountInfoRequest
    ) -> Pin<Box<dyn Future<Output = Result<AccountInfoResponse, SuiError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn handle_object_info_request<'life0, 'async_trait>(
        &'life0 self, 
        request: ObjectInfoRequest
    ) -> Pin<Box<dyn Future<Output = Result<ObjectInfoResponse, SuiError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn handle_transaction_info_request<'life0, 'async_trait>(
        &'life0 self, 
        request: TransactionInfoRequest
    ) -> Pin<Box<dyn Future<Output = Result<TransactionInfoResponse, SuiError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn handle_batch_stream<'life0, 'async_trait>(
        &'life0 self, 
        request: BatchInfoRequest
    ) -> Pin<Box<dyn Future<Output = Result<BatchInfoResponseItemStream, SuiError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Required Methods

fn handle_transaction<'life0, 'async_trait>(
&'life0 self,
transaction: Transaction
) -> Pin<Box<dyn Future<Output = Result<TransactionInfoResponse, SuiError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Initiate a new transaction to a Sui or Primary account.

fn handle_confirmation_transaction<'life0, 'async_trait>(
&'life0 self,
transaction: ConfirmationTransaction
) -> Pin<Box<dyn Future<Output = Result<TransactionInfoResponse, SuiError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Confirm a transaction to a Sui or Primary account.

fn handle_consensus_transaction<'life0, 'async_trait>(
&'life0 self,
transaction: ConsensusTransaction
) -> Pin<Box<dyn Future<Output = Result<TransactionInfoResponse, SuiError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Processes consensus request.

fn handle_account_info_request<'life0, 'async_trait>(
&'life0 self,
request: AccountInfoRequest
) -> Pin<Box<dyn Future<Output = Result<AccountInfoResponse, SuiError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Handle Account information requests for this account.

fn handle_object_info_request<'life0, 'async_trait>(
&'life0 self,
request: ObjectInfoRequest
) -> Pin<Box<dyn Future<Output = Result<ObjectInfoResponse, SuiError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Handle Object information requests for this account.

fn handle_transaction_info_request<'life0, 'async_trait>(
&'life0 self,
request: TransactionInfoRequest
) -> Pin<Box<dyn Future<Output = Result<TransactionInfoResponse, SuiError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Handle Object information requests for this account.

fn handle_batch_stream<'life0, 'async_trait>(
&'life0 self,
request: BatchInfoRequest
) -> Pin<Box<dyn Future<Output = Result<BatchInfoResponseItemStream, SuiError>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,