sui_core

Module checkpoints

Modules

checkpoint_executor

CheckpointExecutor is a Node component that executes all checkpoints for the given epoch. It acts as a Consumer to StateSync for newly synced checkpoints, taking these checkpoints and scheduling and monitoring their execution. Its primary goal is to allow for catching up to the current checkpoint sequence number of the network as quickly as possible so that a newly joined, or recovering Node can participate in a timely manner. To that end, CheckpointExecutor attempts to saturate the CPU with executor tasks (one per checkpoint), each of which handle scheduling and awaiting checkpoint transaction execution.

Structs

BuilderCheckpointSummary

CheckpointAggregator

CheckpointBuilder

CheckpointMetrics

CheckpointRoots

CheckpointService

CheckpointServiceNoop

CheckpointSignatureAggregator

CheckpointStore

CheckpointStoreTables

CheckpointStoreTablesReadOnly

EpochStats

LogCheckpointOutput

PendingCheckpoint

PendingCheckpointInfo

PendingCheckpointV2

PendingCheckpointV2 is merged and split in ConsensusHandler instead of CheckpointBuilder. It contains 1 or more CheckpointRoots which represents a group of transactions settled together. Usage of PendingCheckpointV2 is gated by split_checkpoints_in_consensus_handler We need to support dual implementations until this feature flag is removed.

PollCounter

SendCheckpointToStateSync

SubmitCheckpointToConsensus

Enums

CheckpointBuilderError

CheckpointWatermark

Traits

CheckpointServiceNotify

Type Aliases

CheckpointBuilderResult

CheckpointHeight