Struct sui_core::checkpoints::CheckpointStore

pub struct CheckpointStore { /* private fields */ }

Implementations

impl CheckpointStore

pub fn configurator() -> CheckpointStoreConfigurator

impl CheckpointStore

pub fn open_tables_read_write( path: PathBuf, metric_conf: MetricConf, global_db_options_override: Option<Options>, tables_db_options_override: Option<DBMapTableConfigMap> ) -> Self

Opens a set of tables in read-write mode Only one process is allowed to do this at a time global_db_options_override apply to the whole DB tables_db_options_override apply to each table. If None, the attributes from default_options_override_fn are used if any

pub fn open_tables_transactional( path: PathBuf, metric_conf: MetricConf, global_db_options_override: Option<Options>, tables_db_options_override: Option<DBMapTableConfigMap> ) -> Self

Opens a set of tables in transactional read-write mode Only one process is allowed to do this at a time global_db_options_override apply to the whole DB tables_db_options_override apply to each table. If None, the attributes from default_options_override_fn are used if any

pub fn describe_tables() -> BTreeMap<String, (String, String)>

Returns a list of the tables name and type pairs

pub fn get_read_only_handle( primary_path: PathBuf, with_secondary_path: Option<PathBuf>, global_db_options_override: Option<Options>, metric_conf: MetricConf ) -> CheckpointStoreReadOnly

This opens the DB in read only mode and returns a struct which exposes debug features

impl CheckpointStore

pub fn new(path: &Path) -> Arc<Self>

pub fn open_readonly(path: &Path) -> CheckpointStoreReadOnly

pub fn insert_genesis_checkpoint( &self, checkpoint: VerifiedCheckpoint, contents: CheckpointContents, epoch_store: &AuthorityPerEpochStore )

pub fn get_checkpoint_by_digest( &self, digest: &CheckpointDigest ) -> Result<Option<VerifiedCheckpoint>, TypedStoreError>

pub fn get_checkpoint_by_sequence_number( &self, sequence_number: CheckpointSequenceNumber ) -> Result<Option<VerifiedCheckpoint>, TypedStoreError>

pub fn get_sequence_number_by_contents_digest( &self, digest: &CheckpointContentsDigest ) -> Result<Option<CheckpointSequenceNumber>, TypedStoreError>

pub fn delete_contents_digest_sequence_number_mapping( &self, digest: &CheckpointContentsDigest ) -> Result<(), TypedStoreError>

pub fn get_latest_certified_checkpoint(&self) -> Option<VerifiedCheckpoint>

pub fn multi_get_checkpoint_by_sequence_number( &self, sequence_numbers: &[CheckpointSequenceNumber] ) -> Result<Vec<Option<VerifiedCheckpoint>>, TypedStoreError>

pub fn multi_get_checkpoint_content( &self, contents_digest: &[CheckpointContentsDigest] ) -> Result<Vec<Option<CheckpointContents>>, TypedStoreError>

pub fn get_highest_verified_checkpoint( &self ) -> Result<Option<VerifiedCheckpoint>, TypedStoreError>

pub fn get_highest_synced_checkpoint_seq_number( &self ) -> Result<Option<CheckpointSequenceNumber>, TypedStoreError>

pub fn get_highest_synced_checkpoint( &self ) -> Result<Option<VerifiedCheckpoint>, TypedStoreError>

pub fn get_highest_executed_checkpoint_seq_number( &self ) -> Result<Option<CheckpointSequenceNumber>, TypedStoreError>

pub fn get_highest_executed_checkpoint( &self ) -> Result<Option<VerifiedCheckpoint>, TypedStoreError>

pub fn get_checkpoint_contents( &self, digest: &CheckpointContentsDigest ) -> Result<Option<CheckpointContents>, TypedStoreError>

pub fn get_full_checkpoint_contents_by_sequence_number( &self, seq: CheckpointSequenceNumber ) -> Result<Option<FullCheckpointContents>, TypedStoreError>

pub fn insert_certified_checkpoint( &self, checkpoint: &VerifiedCheckpoint ) -> Result<(), TypedStoreError>

pub fn insert_verified_checkpoint( &self, checkpoint: VerifiedCheckpoint ) -> Result<(), TypedStoreError>

pub fn update_highest_synced_checkpoint( &self, checkpoint: &VerifiedCheckpoint ) -> Result<(), TypedStoreError>

pub fn update_highest_executed_checkpoint( &self, checkpoint: &VerifiedCheckpoint ) -> Result<(), TypedStoreError>

pub fn set_highest_executed_checkpoint_subtle( &self, checkpoint: &VerifiedCheckpoint ) -> Result<(), TypedStoreError>

Sets highest executed checkpoint to any value.

WARNING: This method is very subtle and can corrupt the database if used incorrectly. It should only be used in one-off cases or tests after fully understanding the risk.

pub fn insert_checkpoint_contents( &self, contents: CheckpointContents ) -> Result<(), TypedStoreError>

pub fn insert_verified_checkpoint_contents( &self, checkpoint: &VerifiedCheckpoint, full_contents: VerifiedCheckpointContents ) -> Result<(), TypedStoreError>

pub fn delete_full_checkpoint_contents( &self, seq: CheckpointSequenceNumber ) -> Result<(), TypedStoreError>

pub fn get_epoch_last_checkpoint( &self, epoch_id: EpochId ) -> SuiResult<Option<VerifiedCheckpoint>>

pub fn get_epoch_stats( &self, epoch: EpochId, last_checkpoint: &CheckpointSummary ) -> Option<EpochStats>

Given the epoch ID, and the last checkpoint of the epoch, derive a few statistics of the epoch.

pub fn checkpoint_db(&self, path: &Path) -> SuiResult

pub fn delete_highest_executed_checkpoint_test_only( &self ) -> Result<(), TypedStoreError>

pub fn reset_db_for_execution_since_genesis(&self) -> SuiResult