Trait sui_core::gateway_state::GatewayAPI

pub trait GatewayAPI {

    fn execute_transaction<'life0, 'async_trait>(
        &'life0 self, 
        tx: Transaction
    ) -> Pin<Box<dyn Future<Output = Result<TransactionResponse, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn transfer_coin<'life0, 'async_trait>(
        &'life0 self, 
        signer: SuiAddress, 
        object_id: ObjectID, 
        gas_payment: ObjectID, 
        gas_budget: u64, 
        recipient: SuiAddress
    ) -> Pin<Box<dyn Future<Output = Result<TransactionData, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn sync_account_state<'life0, 'async_trait>(
        &'life0 self, 
        account_addr: SuiAddress
    ) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn move_call<'life0, 'async_trait>(
        &'life0 self, 
        signer: SuiAddress, 
        package_object_ref: ObjectRef, 
        module: Identifier, 
        function: Identifier, 
        type_arguments: Vec<TypeTag>, 
        arguments: Vec<SuiJsonValue>, 
        gas_object_ref: ObjectRef, 
        gas_budget: u64
    ) -> Pin<Box<dyn Future<Output = Result<TransactionData, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn publish<'life0, 'async_trait>(
        &'life0 self, 
        signer: SuiAddress, 
        package_bytes: Vec<Vec<u8>>, 
        gas_object_ref: ObjectRef, 
        gas_budget: u64
    ) -> Pin<Box<dyn Future<Output = Result<TransactionData, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn split_coin<'life0, 'async_trait>(
        &'life0 self, 
        signer: SuiAddress, 
        coin_object_id: ObjectID, 
        split_amounts: Vec<u64>, 
        gas_payment: ObjectID, 
        gas_budget: u64
    ) -> Pin<Box<dyn Future<Output = Result<TransactionData, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn merge_coins<'life0, 'async_trait>(
        &'life0 self, 
        signer: SuiAddress, 
        primary_coin: ObjectID, 
        coin_to_merge: ObjectID, 
        gas_payment: ObjectID, 
        gas_budget: u64
    ) -> Pin<Box<dyn Future<Output = Result<TransactionData, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_object_info<'life0, 'async_trait>(
        &'life0 self, 
        object_id: ObjectID
    ) -> Pin<Box<dyn Future<Output = Result<ObjectRead, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_owned_objects<'life0, 'async_trait>(
        &'life0 self, 
        account_addr: SuiAddress
    ) -> Pin<Box<dyn Future<Output = Result<Vec<ObjectRef>, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn get_total_transaction_number(&self) -> Result<u64, Error>;
    fn get_transactions_in_range(
        &self, 
        start: GatewayTxSeqNumber, 
        end: GatewayTxSeqNumber
    ) -> Result<Vec<(GatewayTxSeqNumber, TransactionDigest)>, Error>;
    fn get_recent_transactions(
        &self, 
        count: u64
    ) -> Result<Vec<(GatewayTxSeqNumber, TransactionDigest)>, Error>;
    fn get_transaction<'life0, 'async_trait>(
        &'life0 self, 
        digest: TransactionDigest
    ) -> Pin<Box<dyn Future<Output = Result<TransactionEffectsResponse, Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

}

Required Methods

fn execute_transaction<'life0, 'async_trait>(
&'life0 self,
tx: Transaction
) -> Pin<Box<dyn Future<Output = Result<TransactionResponse, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

fn transfer_coin<'life0, 'async_trait>(
&'life0 self,
signer: SuiAddress,
object_id: ObjectID,
gas_payment: ObjectID,
gas_budget: u64,
recipient: SuiAddress
) -> Pin<Box<dyn Future<Output = Result<TransactionData, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Send coin object to a Sui address.

fn sync_account_state<'life0, 'async_trait>(
&'life0 self,
account_addr: SuiAddress
) -> Pin<Box<dyn Future<Output = Result<(), Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Synchronise account state with a random authorities, updates all object_ids from account_addr, request only goes out to one authority. this method doesn’t guarantee data correctness, caller will have to handle potential byzantine authority

fn move_call<'life0, 'async_trait>(
&'life0 self,
signer: SuiAddress,
package_object_ref: ObjectRef,
module: Identifier,
function: Identifier,
type_arguments: Vec<TypeTag>,
arguments: Vec<SuiJsonValue>,
gas_object_ref: ObjectRef,
gas_budget: u64
) -> Pin<Box<dyn Future<Output = Result<TransactionData, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Call move functions in the module in the given package, with args supplied

fn publish<'life0, 'async_trait>(
&'life0 self,
signer: SuiAddress,
package_bytes: Vec<Vec<u8>>,
gas_object_ref: ObjectRef,
gas_budget: u64
) -> Pin<Box<dyn Future<Output = Result<TransactionData, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Publish Move modules

fn split_coin<'life0, 'async_trait>(
&'life0 self,
signer: SuiAddress,
coin_object_id: ObjectID,
split_amounts: Vec<u64>,
gas_payment: ObjectID,
gas_budget: u64
) -> Pin<Box<dyn Future<Output = Result<TransactionData, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Split the coin object (identified by coin_object_ref) into multiple new coins. The amount of each new coin is specified in split_amounts. Remaining balance is kept in the original coin object. Note that the order of the new coins in SplitCoinResponse will not be the same as the order of split_amounts.

fn merge_coins<'life0, 'async_trait>(
&'life0 self,
signer: SuiAddress,
primary_coin: ObjectID,
coin_to_merge: ObjectID,
gas_payment: ObjectID,
gas_budget: u64
) -> Pin<Box<dyn Future<Output = Result<TransactionData, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Merge the coin_to_merge coin object into primary_coin. After this merge, the balance of primary_coin will become the sum of the two, while coin_to_merge will be deleted.

Returns a pair: (update primary coin object reference, updated gas payment object reference)

TODO: Support merging a vector of coins.

fn get_object_info<'life0, 'async_trait>(
&'life0 self,
object_id: ObjectID
) -> Pin<Box<dyn Future<Output = Result<ObjectRead, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get the object information

fn get_owned_objects<'life0, 'async_trait>(
&'life0 self,
account_addr: SuiAddress
) -> Pin<Box<dyn Future<Output = Result<Vec<ObjectRef>, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Get refs of all objects we own from local cache.

fn get_total_transaction_number(&self) -> Result<u64, Error>

Get the total number of transactions ever happened in history.

fn get_transactions_in_range(
&self,
start: GatewayTxSeqNumber,
end: GatewayTxSeqNumber
) -> Result<Vec<(GatewayTxSeqNumber, TransactionDigest)>, Error>

Return the list of transactions with sequence number in range [start, end). start is included, end is excluded.

fn get_recent_transactions(
&self,
count: u64
) -> Result<Vec<(GatewayTxSeqNumber, TransactionDigest)>, Error>

Return the most recent count transactions.

fn get_transaction<'life0, 'async_trait>(
&'life0 self,
digest: TransactionDigest
) -> Pin<Box<dyn Future<Output = Result<TransactionEffectsResponse, Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,