sui_core::jsonrpc_index
Struct IndexStore

pub struct IndexStore {
    pub caches: IndexStoreCaches,
    /* private fields */
}

Fields
caches: IndexStoreCaches
Implementations

impl IndexStore

pub fn new_without_init( path: PathBuf, registry: &Registry, max_type_length: Option<u64>, remove_deprecated_tables: bool, ) -> Self

pub fn new( path: PathBuf, registry: &Registry, max_type_length: Option<u64>, remove_deprecated_tables: bool, ) -> Self

pub fn tables(&self) -> &IndexStoreTables

pub fn index_coin( &self, digest: &TransactionDigest, batch: &mut DBBatch, object_index_changes: &ObjectIndexChanges, tx_coins: Option<TxCoins>, ) -> SuiResult<IndexStoreCacheUpdates>

pub fn index_tx( &self, sender: SuiAddress, active_inputs: impl Iterator<Item = ObjectID>, mutated_objects: impl Iterator<Item = (ObjectRef, Owner)> + Clone, move_functions: impl Iterator<Item = (ObjectID, String, String)> + Clone, events: &TransactionEvents, object_index_changes: ObjectIndexChanges, digest: &TransactionDigest, timestamp_ms: u64, tx_coins: Option<TxCoins>, ) -> SuiResult<u64>

pub fn next_sequence_number(&self) -> TxSequenceNumber

pub fn get_transactions( &self, filter: Option<TransactionFilter>, cursor: Option<TransactionDigest>, limit: Option<usize>, reverse: bool, ) -> SuiResult<Vec<TransactionDigest>>

pub fn get_transactions_by_input_object( &self, input_object: ObjectID, cursor: Option<TxSequenceNumber>, limit: Option<usize>, reverse: bool, ) -> SuiResult<Vec<TransactionDigest>>

pub fn get_transactions_by_mutated_object( &self, mutated_object: ObjectID, cursor: Option<TxSequenceNumber>, limit: Option<usize>, reverse: bool, ) -> SuiResult<Vec<TransactionDigest>>

pub fn get_transactions_from_addr( &self, addr: SuiAddress, cursor: Option<TxSequenceNumber>, limit: Option<usize>, reverse: bool, ) -> SuiResult<Vec<TransactionDigest>>

pub fn get_transactions_by_move_function( &self, package: ObjectID, module: Option<String>, function: Option<String>, cursor: Option<TxSequenceNumber>, limit: Option<usize>, reverse: bool, ) -> SuiResult<Vec<TransactionDigest>>

pub fn get_transactions_to_addr( &self, addr: SuiAddress, cursor: Option<TxSequenceNumber>, limit: Option<usize>, reverse: bool, ) -> SuiResult<Vec<TransactionDigest>>

pub fn get_transaction_seq( &self, digest: &TransactionDigest, ) -> SuiResult<Option<TxSequenceNumber>>

pub fn all_events( &self, tx_seq: TxSequenceNumber, event_seq: usize, limit: usize, descending: bool, ) -> SuiResult<Vec<(TransactionEventsDigest, TransactionDigest, usize, u64)>>

pub fn events_by_transaction( &self, digest: &TransactionDigest, tx_seq: TxSequenceNumber, event_seq: usize, limit: usize, descending: bool, ) -> SuiResult<Vec<(TransactionEventsDigest, TransactionDigest, usize, u64)>>

pub fn events_by_module_id( &self, module: &ModuleId, tx_seq: TxSequenceNumber, event_seq: usize, limit: usize, descending: bool, ) -> SuiResult<Vec<(TransactionEventsDigest, TransactionDigest, usize, u64)>>

pub fn events_by_move_event_struct_name( &self, struct_name: &StructTag, tx_seq: TxSequenceNumber, event_seq: usize, limit: usize, descending: bool, ) -> SuiResult<Vec<(TransactionEventsDigest, TransactionDigest, usize, u64)>>

pub fn events_by_move_event_module( &self, module_id: &ModuleId, tx_seq: TxSequenceNumber, event_seq: usize, limit: usize, descending: bool, ) -> SuiResult<Vec<(TransactionEventsDigest, TransactionDigest, usize, u64)>>

pub fn events_by_sender( &self, sender: &SuiAddress, tx_seq: TxSequenceNumber, event_seq: usize, limit: usize, descending: bool, ) -> SuiResult<Vec<(TransactionEventsDigest, TransactionDigest, usize, u64)>>

pub fn event_iterator( &self, start_time: u64, end_time: u64, tx_seq: TxSequenceNumber, event_seq: usize, limit: usize, descending: bool, ) -> SuiResult<Vec<(TransactionEventsDigest, TransactionDigest, usize, u64)>>

pub fn prune(&self, cut_time_ms: u64) -> SuiResult<TxSequenceNumber>

pub fn get_dynamic_fields_iterator( &self, object: ObjectID, cursor: Option<ObjectID>, ) -> SuiResult<impl Iterator<Item = Result<(ObjectID, DynamicFieldInfo), TypedStoreError>> + '_>

pub fn get_dynamic_field_object_id( &self, object: ObjectID, name_type: TypeTag, name_bcs_bytes: &[u8], ) -> SuiResult<Option<ObjectID>>

pub fn get_owner_objects( &self, owner: SuiAddress, cursor: Option<ObjectID>, limit: usize, filter: Option<SuiObjectDataFilter>, ) -> SuiResult<Vec<ObjectInfo>>

pub fn get_owned_coins_iterator( coin_index: &DBMap<CoinIndexKey2, CoinInfo>, owner: SuiAddress, coin_type_tag: Option<String>, ) -> SuiResult<impl Iterator<Item = (CoinIndexKey2, CoinInfo)> + '_>

pub fn get_owned_coins_iterator_with_cursor( &self, owner: SuiAddress, cursor: (String, u64, ObjectID), limit: usize, one_coin_type_only: bool, ) -> SuiResult<impl Iterator<Item = (CoinIndexKey2, CoinInfo)> + '_>

pub fn get_owner_objects_iterator( &self, owner: SuiAddress, starting_object_id: ObjectID, filter: Option<SuiObjectDataFilter>, ) -> SuiResult<impl Iterator<Item = ObjectInfo> + '_>

starting_object_id can be used to implement pagination, where a client remembers the last object id of each page, and use it to query the next page.

pub fn insert_genesis_objects( &self, object_index_changes: ObjectIndexChanges, ) -> SuiResult

pub fn is_empty(&self) -> bool

pub fn checkpoint_db(&self, path: &Path) -> SuiResult

pub fn get_balance( &self, owner: SuiAddress, coin_type: TypeTag, ) -> SuiResult<TotalBalance>

This method first gets the balance from per_coin_type_balance cache. On a cache miss, it gets the balance for passed in coin_type from the all_balance cache. Only on the second cache miss, we go to the database (expensive) and update the cache. Notice that db read is done with spawn_blocking as that is expected to block

pub fn get_all_balance( &self, owner: SuiAddress, ) -> SuiResult<Arc<HashMap<TypeTag, TotalBalance>>>

This method gets the balance for all coin types from the all_balance cache. On a cache miss, we go to the database (expensive) and update the cache. This cache is dual purpose in the sense that it not only serves get_AllBalance() calls but is also used for serving get_Balance() queries. Notice that db read is performed with spawn_blocking as that is expected to block

pub fn get_balance_from_db( metrics: Arc<IndexStoreMetrics>, coin_index: DBMap<CoinIndexKey2, CoinInfo>, owner: SuiAddress, coin_type: TypeTag, ) -> SuiResult<TotalBalance>

Read balance for a SuiAddress and CoinType from the backend database

pub fn get_all_balances_from_db( metrics: Arc<IndexStoreMetrics>, coin_index: DBMap<CoinIndexKey2, CoinInfo>, owner: SuiAddress, ) -> SuiResult<Arc<HashMap<TypeTag, TotalBalance>>>

Read all balances for a SuiAddress from the backend database