Module sui_core::signature_verifier

Structs

SignatureVerifier

Verifies signatures in ways that faster than verifying each signature individually.

SignatureVerifierMetrics

Functions

batch_verify_all_certificates_and_checkpoints

Verifies all certificates - if any fail return error.

batch_verify_certificates

Verifies certificates in batch mode, but returns a separate result for each cert.