Trait sui_core::state_accumulator::AccumulatorReadStore

pub trait AccumulatorReadStore {
    // Required methods
    fn multi_get_object_by_key(
        &self,
        object_keys: &[ObjectKey]
    ) -> SuiResult<Vec<Option<Object>>>;
    fn get_object_ref_prior_to_key_deprecated(
        &self,
        object_id: &ObjectID,
        version: VersionNumber
    ) -> SuiResult<Option<ObjectRef>>;
}

Required Methods

fn multi_get_object_by_key( &self, object_keys: &[ObjectKey] ) -> SuiResult<Vec<Option<Object>>>

fn get_object_ref_prior_to_key_deprecated( &self, object_id: &ObjectID, version: VersionNumber ) -> SuiResult<Option<ObjectRef>>

This function is only called in older protocol versions, and should no longer be used. It creates an explicit dependency to tombstones which is not desired.