sui_core
Module validator_tx_finalizer

Structs

ValidatorTxFinalizer

The ValidatorTxFinalizer is responsible for finalizing transactions that have been signed by the validator. It does this by waiting for a delay after the transaction has been signed, and then attempting to finalize the transaction if it has not yet been done by a fullnode.

ValidatorTxFinalizerConfig