Crate sui_core

Modules

authority

authority_aggregator

authority_client

authority_server

checkpoints

congestion_tracker

consensus_adapter

consensus_handler

consensus_manager

consensus_throughput_calculator

consensus_validator

db_checkpoint_handler

epoch

execution_cache

jsonrpc_index

IndexStore supports creation of various ancillary indexes of state in SuiDataStore. The main user of this data is the explorer.

metrics

mock_consensus

module_cache_metrics

mysticeti_adapter

overload_monitor

quorum_driver

rpc_index

runtime

safe_client

signature_verifier

state_accumulator

storage

streamer

subscription_handler

test_authority_clients

test_utils

traffic_controller

transaction_orchestrator

validator_tx_finalizer

verify_indexes

Macros

handle_with_decoration

Implements generic pre- and post-processing. Since this is on the critical path, any heavy lifting should be done in a separate non-blocking task unless it is necessary to override the return value.