sui_core::authority::authority_per_epoch_store
Struct AuthorityPerEpochStore

pub struct AuthorityPerEpochStore { /* private fields */ }

Implementations

impl AuthorityPerEpochStore

pub fn new( name: AuthorityName, committee: Arc<Committee>, parent_path: &Path, db_options: Option<Options>, metrics: Arc<EpochMetrics>, epoch_start_configuration: EpochStartConfiguration, backing_package_store: Arc<dyn BackingPackageStore + Send + Sync>, object_store: Arc<dyn ObjectStore + Send + Sync>, cache_metrics: Arc<ResolverMetrics>, signature_verifier_metrics: Arc<SignatureVerifierMetrics>, expensive_safety_check_config: &ExpensiveSafetyCheckConfig, chain: (ChainIdentifier, Chain), highest_executed_checkpoint: CheckpointSequenceNumber, ) -> SuiResult<Arc<Self>>

pub fn tables(&self) -> SuiResult<Arc<AuthorityEpochTables>>

pub fn release_db_handles(&self)

pub fn authenticator_state_enabled(&self) -> bool

pub fn authenticator_state_exists(&self) -> bool

pub fn randomness_state_enabled(&self) -> bool

pub fn randomness_state_exists(&self) -> bool

pub fn randomness_reporter(&self) -> Option<RandomnessReporter>

pub async fn set_randomness_manager( &self, randomness_manager: RandomnessManager, ) -> SuiResult<()>

pub fn accumulator_root_exists(&self) -> bool

pub fn accumulators_enabled(&self) -> bool

pub fn coin_deny_list_state_exists(&self) -> bool

pub fn coin_deny_list_v1_enabled(&self) -> bool

pub fn bridge_exists(&self) -> bool

pub fn bridge_committee_initiated(&self) -> bool

pub fn get_parent_path(&self) -> PathBuf

pub fn epoch_start_config(&self) -> &Arc<EpochStartConfiguration>

Returns &Arc<EpochStartConfiguration> User can treat this Arc as &EpochStartConfiguration, or clone the Arc to pass as owned object

pub fn epoch_start_state(&self) -> &EpochStartSystemState

pub fn get_chain_identifier(&self) -> ChainIdentifier

pub fn get_chain(&self) -> Chain

pub fn new_at_next_epoch( &self, name: AuthorityName, new_committee: Committee, epoch_start_configuration: EpochStartConfiguration, backing_package_store: Arc<dyn BackingPackageStore + Send + Sync>, object_store: Arc<dyn ObjectStore + Send + Sync>, expensive_safety_check_config: &ExpensiveSafetyCheckConfig, previous_epoch_last_checkpoint: CheckpointSequenceNumber, ) -> SuiResult<Arc<Self>>

pub fn new_at_next_epoch_for_testing( &self, backing_package_store: Arc<dyn BackingPackageStore + Send + Sync>, object_store: Arc<dyn ObjectStore + Send + Sync>, expensive_safety_check_config: &ExpensiveSafetyCheckConfig, previous_epoch_last_checkpoint: CheckpointSequenceNumber, ) -> Arc<Self>

pub fn committee(&self) -> &Arc<Committee>

pub fn protocol_config(&self) -> &ProtocolConfig

pub fn epoch(&self) -> EpochId

pub fn get_state_hash_for_checkpoint( &self, checkpoint: &CheckpointSequenceNumber, ) -> SuiResult<Option<GlobalStateHash>>

pub fn insert_state_hash_for_checkpoint( &self, checkpoint: &CheckpointSequenceNumber, accumulator: &GlobalStateHash, ) -> SuiResult

pub fn get_running_root_state_hash( &self, checkpoint: CheckpointSequenceNumber, ) -> SuiResult<Option<GlobalStateHash>>

pub fn get_highest_running_root_state_hash( &self, ) -> SuiResult<Option<(CheckpointSequenceNumber, GlobalStateHash)>>

pub fn insert_running_root_state_hash( &self, checkpoint: &CheckpointSequenceNumber, hash: &GlobalStateHash, ) -> SuiResult

pub fn reference_gas_price(&self) -> u64

pub fn protocol_version(&self) -> ProtocolVersion

pub fn module_cache(&self) -> &Arc<SyncModuleCache<ResolverWrapper>>

pub fn executor(&self) -> &Arc<dyn Executor + Send + Sync>

pub fn set_local_execution_time_channels( &self, tx_local_execution_time: Sender<(ProgrammableTransaction, Vec<ExecutionTiming>, Duration)>, tx_object_debts: Sender<Vec<ObjectID>>, )

pub fn record_local_execution_time( &self, tx: &TransactionData, effects: &TransactionEffects, timings: Vec<ExecutionTiming>, total_duration: Duration, )

pub fn get_stored_execution_time_observations( protocol_config: &ProtocolConfig, committee: Arc<Committee>, object_store: &dyn ObjectStore, metrics: &EpochMetrics, ) -> impl Iterator<Item = (AuthorityIndex, u64, ExecutionTimeObservationKey, Duration)>

pub fn get_end_of_epoch_execution_time_observations( &self, ) -> &StoredExecutionTimeObservations

pub fn acquire_tx_guard( &self, cert: &VerifiedExecutableTransaction, ) -> SuiResult<CertTxGuard>

pub fn acquire_tx_lock(&self, digest: &TransactionDigest) -> CertLockGuard

Acquire the lock for a tx without writing to the WAL.

pub fn store_reconfig_state(&self, new_state: &ReconfigState) -> SuiResult

pub fn insert_signed_transaction( &self, transaction: VerifiedSignedTransaction, ) -> SuiResult

pub fn get_signed_transaction( &self, tx_digest: &TransactionDigest, ) -> SuiResult<Option<VerifiedSignedTransaction>>

pub fn insert_tx_cert_sig( &self, tx_digest: &TransactionDigest, cert_sig: &AuthorityStrongQuorumSignInfo, ) -> SuiResult

pub fn insert_tx_key( &self, tx_key: &TransactionKey, tx_digest: &TransactionDigest, ) -> SuiResult

pub fn num_shared_version_assignments(&self) -> usize

pub fn revert_executed_transaction( &self, tx_digest: &TransactionDigest, ) -> SuiResult

pub fn insert_effects_digest_and_signature( &self, tx_digest: &TransactionDigest, effects_digest: &TransactionEffectsDigest, effects_signature: &AuthoritySignInfo, ) -> SuiResult

pub fn transactions_executed_in_cur_epoch( &self, digests: &[TransactionDigest], ) -> SuiResult<Vec<bool>>

pub fn get_effects_signature( &self, tx_digest: &TransactionDigest, ) -> SuiResult<Option<AuthoritySignInfo>>

pub fn get_signed_effects_digest( &self, tx_digest: &TransactionDigest, ) -> SuiResult<Option<TransactionEffectsDigest>>

pub fn get_transaction_cert_sig( &self, tx_digest: &TransactionDigest, ) -> SuiResult<Option<AuthorityStrongQuorumSignInfo>>

pub fn get_last_consensus_stats(&self) -> SuiResult<ExecutionIndicesWithStats>

pub fn get_accumulators_in_checkpoint_range( &self, from_checkpoint: CheckpointSequenceNumber, to_checkpoint: CheckpointSequenceNumber, ) -> SuiResult<Vec<(CheckpointSequenceNumber, GlobalStateHash)>>

pub async fn notify_read_checkpoint_state_hasher( &self, checkpoints: &[CheckpointSequenceNumber], ) -> SuiResult<Vec<GlobalStateHash>>

Returns future containing the state accumulator for the given epoch once available.

pub async fn notify_read_running_root( &self, checkpoint: CheckpointSequenceNumber, ) -> SuiResult<GlobalStateHash>

pub fn handle_finalized_checkpoint( &self, checkpoint: &CheckpointSummary, digests: &[TransactionDigest], ) -> SuiResult<()>

Called when transaction outputs are committed to disk

pub fn get_all_pending_consensus_transactions( &self, ) -> Vec<ConsensusTransaction>

pub fn set_shared_object_versions_for_testing( &self, tx_digest: &TransactionDigest, assigned_versions: &[(ConsensusObjectSequenceKey, SequenceNumber)], ) -> SuiResult

pub fn insert_finalized_transactions( &self, digests: &[TransactionDigest], sequence: CheckpointSequenceNumber, ) -> SuiResult

pub fn is_transaction_executed_in_checkpoint( &self, digest: &TransactionDigest, ) -> SuiResult<bool>

pub fn transactions_executed_in_checkpoint( &self, digests: impl Iterator<Item = TransactionDigest>, ) -> SuiResult<Vec<bool>>

pub fn get_transaction_checkpoint( &self, digest: &TransactionDigest, ) -> SuiResult<Option<CheckpointSequenceNumber>>

pub fn multi_get_transaction_checkpoint( &self, digests: &[TransactionDigest], ) -> SuiResult<Vec<Option<CheckpointSequenceNumber>>>

pub fn get_assigned_shared_object_versions( &self, key: &TransactionKey, ) -> Option<Vec<(ConsensusObjectSequenceKey, SequenceNumber)>>

pub fn assign_shared_object_versions_idempotent( &self, cache_reader: &dyn ObjectCacheRead, certificates: &[VerifiedExecutableTransaction], ) -> SuiResult

Given list of certificates, assign versions for all shared objects used in them. We start with the current next_shared_object_versions table for each object, and build up the versions based on the dependencies of each certificate. However, in the end we do not update the next_shared_object_versions table, which keeps this function idempotent. We should call this function when we are assigning shared object versions outside of consensus and do not want to taint the next_shared_object_versions table.

pub fn get_all_deferred_transactions_for_test( &self, ) -> Vec<(DeferralKey, Vec<VerifiedSequencedConsensusTransaction>)>

pub fn acquire_shared_version_assignments_from_effects( &self, certificate: &VerifiedExecutableTransaction, effects: &TransactionEffects, cache_reader: &dyn ObjectCacheRead, ) -> SuiResult

Assign a sequence number for the shared objects of the input transaction based on the effects of that transaction. Used by full nodes who donโt listen to consensus, and validators who catch up by state sync.

pub fn insert_pending_consensus_transactions( &self, transactions: &[ConsensusTransaction], lock: Option<&RwLockReadGuard<'_, ReconfigState>>, ) -> SuiResult

When submitting a certificate caller must provide a ReconfigState lock guard and verify that it allows new user certificates

pub fn remove_pending_consensus_transactions( &self, keys: &[ConsensusTransactionKey], ) -> SuiResult

pub fn pending_consensus_certificates_count(&self) -> usize

pub fn pending_consensus_certificates_empty(&self) -> bool

pub fn pending_consensus_certificates(&self) -> HashSet<TransactionDigest>

pub fn is_pending_consensus_certificate( &self, tx_digest: &TransactionDigest, ) -> bool

pub fn deferred_transactions_empty(&self) -> bool

pub fn is_any_tx_certs_consensus_message_processed<'a>( &self, certificates: impl Iterator<Item = &'a CertifiedTransaction>, ) -> SuiResult<bool>

Check whether any certificates were processed by consensus. This handles multiple certificates at once.

pub fn all_external_consensus_messages_processed( &self, keys: impl Iterator<Item = ConsensusTransactionKey>, ) -> SuiResult<bool>

Returns true if all messages with the given keys were processed by consensus.

pub fn is_consensus_message_processed( &self, key: &SequencedConsensusTransactionKey, ) -> SuiResult<bool>

pub fn check_consensus_messages_processed( &self, keys: impl Iterator<Item = SequencedConsensusTransactionKey>, ) -> SuiResult<Vec<bool>>

pub async fn consensus_messages_processed_notify( &self, keys: Vec<SequencedConsensusTransactionKey>, ) -> Result<(), SuiError>

pub async fn transactions_executed_in_checkpoint_notify( &self, digests: Vec<TransactionDigest>, ) -> Result<(), SuiError>

Get notified when transactions get executed as part of a checkpoint execution.

pub fn has_sent_end_of_publish( &self, authority: &AuthorityName, ) -> SuiResult<bool>

pub async fn notify_read_executed_digests( &self, keys: &[TransactionKey], ) -> SuiResult<Vec<TransactionDigest>>

pub fn user_signatures_for_checkpoint( &self, transactions: &[VerifiedTransaction], digests: &[TransactionDigest], ) -> SuiResult<Vec<Vec<GenericSignature>>>

Note: caller usually need to call consensus_message_processed_notify before this call

pub fn clear_override_protocol_upgrade_buffer_stake(&self) -> SuiResult

pub fn set_override_protocol_upgrade_buffer_stake( &self, new_stake_bps: u64, ) -> SuiResult

pub fn get_effective_buffer_stake_bps(&self) -> u64

pub fn record_capabilities( &self, capabilities: &AuthorityCapabilitiesV1, ) -> SuiResult

Record most recently advertised capabilities of all authorities

pub fn record_capabilities_v2( &self, capabilities: &AuthorityCapabilitiesV2, ) -> SuiResult

Record most recently advertised capabilities of all authorities

pub fn get_capabilities_v1(&self) -> SuiResult<Vec<AuthorityCapabilitiesV1>>

pub fn get_capabilities_v2(&self) -> SuiResult<Vec<AuthorityCapabilitiesV2>>

pub fn jwk_active_in_current_epoch(&self, jwk_id: &JwkId, jwk: &JWK) -> bool

pub fn get_reconfig_state_read_lock_guard( &self, ) -> RwLockReadGuard<'_, ReconfigState>

pub fn get_reconfig_state_write_lock_guard( &self, ) -> RwLockWriteGuard<'_, ReconfigState>

pub fn close_user_certs(&self, lock_guard: RwLockWriteGuard<'_, ReconfigState>)

pub async fn user_certs_closed_notify(&self)

pub async fn epoch_terminated(&self)

Notify epoch is terminated, can only be called once on epoch store

pub async fn wait_epoch_terminated(&self)

Waits for the notification about epoch termination

pub async fn within_alive_epoch<F: Future + Send>( &self, f: F, ) -> Result<F::Output, ()>

This function executes given future until epoch_terminated is called If future finishes before epoch_terminated is called, future result is returned If epoch_terminated is called before future is resolved, error is returned

In addition to the early termination guarantee, this function also prevents epoch_terminated() if future is being executed.

pub fn verify_transaction( &self, tx: Transaction, ) -> SuiResult<VerifiedTransaction>

pub fn get_highest_pending_checkpoint_height(&self) -> CheckpointHeight

pub async fn process_consensus_transactions_for_tests<C: CheckpointServiceNotify>( self: &Arc<Self>, transactions: Vec<SequencedConsensusTransaction>, checkpoint_service: &Arc<C>, cache_reader: &dyn ObjectCacheRead, tx_reader: &dyn TransactionCacheRead, authority_metrics: &Arc<AuthorityMetrics>, skip_consensus_commit_prologue_in_test: bool, ) -> SuiResult<Vec<VerifiedExecutableTransaction>>

pub fn assign_shared_object_versions_for_tests( self: &Arc<Self>, cache_reader: &dyn ObjectCacheRead, transactions: &[VerifiedExecutableTransaction], ) -> SuiResult

pub fn get_pending_checkpoints( &self, last: Option<CheckpointHeight>, ) -> SuiResult<Vec<(CheckpointHeight, PendingCheckpointV2)>>

pub fn pending_checkpoint_exists( &self, index: &CheckpointHeight, ) -> SuiResult<bool>

pub fn process_constructed_checkpoint( &self, commit_height: CheckpointHeight, content_info: NonEmpty<(CheckpointSummary, CheckpointContents)>, )

pub fn put_genesis_checkpoint_in_builder( &self, summary: &CheckpointSummary, contents: &CheckpointContents, ) -> SuiResult<()>

Register genesis checkpoint in builder DB

pub fn last_built_checkpoint_builder_summary( &self, ) -> SuiResult<Option<BuilderCheckpointSummary>>

pub fn last_built_checkpoint_summary( &self, ) -> SuiResult<Option<(CheckpointSequenceNumber, CheckpointSummary)>>

pub fn get_built_checkpoint_summary( &self, sequence: CheckpointSequenceNumber, ) -> SuiResult<Option<CheckpointSummary>>

pub fn builder_included_transactions_in_checkpoint<'a>( &self, digests: impl Iterator<Item = &'a TransactionDigest>, ) -> SuiResult<Vec<bool>>

pub fn get_last_checkpoint_signature_index(&self) -> SuiResult<u64>

pub fn insert_checkpoint_signature( &self, checkpoint_seq: CheckpointSequenceNumber, index: u64, info: &CheckpointSignatureMessage, ) -> SuiResult<()>

pub fn record_end_of_message_quorum_time_metric(&self)

pub fn record_epoch_reconfig_start_time_metric(&self)

pub fn record_is_safe_mode_metric(&self, safe_mode: bool)

pub fn record_checkpoint_builder_is_safe_mode_metric(&self, safe_mode: bool)

pub fn clear_signature_cache(&self)

pub async fn get_estimated_tx_cost(&self, tx: &TransactionData) -> Option<u64>

Only used by admin API

pub async fn get_consensus_tx_cost_estimates( &self, ) -> Vec<(ExecutionTimeObservationKey, ConsensusObservations)>