Struct sui_core::authority::AuthorityState

pub struct AuthorityState {
    pub name: AuthorityName,
    pub secret: StableSyncAuthoritySigner,
    pub database: Arc<AuthorityStore>,
    pub module_cache: Arc<SyncModuleCache<ResolverWrapper<AuthorityStore>>>,
    pub event_handler: Option<Arc<EventHandler>>,
    pub transaction_streamer: Option<Arc<TransactionStreamer>>,
    pub metrics: Arc<AuthorityMetrics>,
    /* private fields */
}

Fields
name: AuthorityName

The name of this authority.

secret: StableSyncAuthoritySigner

The signature key of the authority.

database: Arc<AuthorityStore>

The database

module_cache: Arc<SyncModuleCache<ResolverWrapper<AuthorityStore>>>event_handler: Option<Arc<EventHandler>>transaction_streamer: Option<Arc<TransactionStreamer>>metrics: Arc<AuthorityMetrics>
Implementations

impl AuthorityState

The authority state encapsulates all state, drives execution, and ensures safety.

Note the authority operations can be accessed through a read ref (&) and do not require &mut. Internally a database is synchronized through a mutex lock.

Repeating valid commands should produce no changes and return no error.

pub fn is_validator(&self) -> bool

pub fn is_fullnode(&self) -> bool

pub fn epoch(&self) -> EpochId

pub fn committee_store(&self) -> &Arc<CommitteeStore>

pub fn clone_committee_store(&self) -> Arc<CommitteeStore>

pub async fn handle_transaction(
&self,
transaction: VerifiedTransaction
) -> Result<VerifiedTransactionInfoResponse, SuiError>

Initiate a new transaction.

pub async fn fullnode_execute_certificate_with_effects(
&self,
certificate: &VerifiedCertificate,
effects: &VerifiedCertifiedTransactionEffects,
epoch_store: &Arc<AuthorityPerEpochStore>
) -> SuiResult

Executes a certificate that’s known to have correct effects. For such certificate, we don’t have to wait for consensus to set shared object locks because we already know the shared object versions based on the effects. This function can be called by a fullnode only.

pub async fn execute_certificate(
&self,
certificate: &VerifiedCertificate,
epoch_store: &Arc<AuthorityPerEpochStore>
) -> SuiResult<VerifiedSignedTransactionEffects>

Executes a certificate for its effects.

pub async fn try_execute_for_test(
&self,
certificate: &VerifiedCertificate
) -> SuiResult<VerifiedSignedTransactionEffects>

Test only wrapper for try_execute_immediately() above, useful for checking errors if the pre-conditions are not satisfied, and executing change epoch transactions.

pub async fn notify_read_effects(
&self,
certificate: &VerifiedCertificate
) -> SuiResult<VerifiedSignedTransactionEffects>

pub fn certificate_executed(
&self,
digest: &TransactionDigest,
epoch_store: &Arc<AuthorityPerEpochStore>
)

Notifies TransactionManager about an executed certificate.

pub async fn dry_exec_transaction(
&self,
transaction: TransactionData,
transaction_digest: TransactionDigest
) -> Result<SuiTransactionEffects, Error>

pub async fn dev_inspect_transaction(
&self,
sender: SuiAddress,
transaction_kind: TransactionKind,
gas_price: u64,
epoch: EpochId
) -> Result<DevInspectResults, Error>

The object ID for gas can be any object ID, even for an uncreated object

pub fn is_tx_already_executed(
&self,
digest: &TransactionDigest
) -> SuiResult<bool>

pub fn unixtime_now_ms() -> u64

pub async fn handle_transaction_info_request(
&self,
request: TransactionInfoRequest
) -> Result<VerifiedTransactionInfoResponse, SuiError>

pub async fn handle_object_info_request(
&self,
request: ObjectInfoRequest
) -> Result<VerifiedObjectInfoResponse, SuiError>

pub fn handle_checkpoint_request(
&self,
request: &CheckpointRequest
) -> Result<CheckpointResponse, SuiError>

pub fn handle_committee_info_request(
&self,
request: &CommitteeInfoRequest
) -> SuiResult<CommitteeInfoResponse>

pub async fn new(
name: AuthorityName,
secret: StableSyncAuthoritySigner,
store: Arc<AuthorityStore>,
epoch_store: Arc<AuthorityPerEpochStore>,
committee_store: Arc<CommitteeStore>,
indexes: Option<Arc<IndexStore>>,
event_store: Option<Arc<EventStoreType>>,
transaction_streamer: Option<Arc<TransactionStreamer>>,
checkpoint_store: Arc<CheckpointStore>,
prometheus_registry: &Registry,
pruning_config: &AuthorityStorePruningConfig
) -> Arc<Self>

pub async fn new_for_testing(
genesis_committee: Committee,
key: &AuthorityKeyPair,
store_base_path: Option<PathBuf>,
genesis: &Genesis
) -> Arc<Self>

pub fn transaction_manager(&self) -> &Arc<TransactionManager>

pub fn enqueue_certificates_for_execution(
&self,
certs: Vec<VerifiedCertificate>,
epoch_store: &Arc<AuthorityPerEpochStore>
) -> SuiResult<()>

Adds certificates to the pending certificate store and transaction manager for ordered execution.

pub async fn process_tx_recovery_log(
&self,
limit: Option<usize>,
epoch_store: &Arc<AuthorityPerEpochStore>
) -> SuiResult

pub async fn reconfigure(
&self,
new_committee: Committee,
epoch_start_timestamp_ms: u64
) -> SuiResult

pub fn db(&self) -> Arc<AuthorityStore>

pub fn epoch_store(&self) -> Guard<Arc<AuthorityPerEpochStore>>

pub fn epoch_store_for_testing(&self) -> Guard<Arc<AuthorityPerEpochStore>>

pub fn load_epoch_store(
&self,
intended_epoch: EpochId
) -> SuiResult<Guard<Arc<AuthorityPerEpochStore>>>

pub fn clone_committee_for_testing(&self) -> Committee

pub async fn get_framework_object_ref(&self) -> SuiResult<ObjectRef>

pub fn get_sui_system_state_object(&self) -> SuiResult<SuiSystemState>

pub async fn get_object_read(
&self,
object_id: &ObjectID
) -> Result<ObjectRead, SuiError>

pub async fn get_past_object_read(
&self,
object_id: &ObjectID,
version: SequenceNumber
) -> Result<PastObjectRead, SuiError>

This function aims to serve rpc reads on past objects and we don’t expect it to be called for other purposes. Depending on the object pruning policies that will be enforced in the future there is no software-level guarantee/SLA to retrieve an object with an old version even if it exists/existed.

pub fn get_owner_objects(&self, owner: SuiAddress) -> SuiResult<Vec<ObjectInfo>>

pub fn get_owner_objects_iterator(
&self,
owner: SuiAddress
) -> SuiResult<impl Iterator<Item = ObjectInfo> + '_>

pub async fn get_move_objects<T>(
&self,
owner: SuiAddress,
type_: &StructTag
) -> SuiResult<Vec<T>>where
T: DeserializeOwned,

pub fn get_dynamic_fields(
&self,
owner: ObjectID,
cursor: Option<ObjectID>,
limit: usize
) -> SuiResult<Vec<DynamicFieldInfo>>

pub fn get_dynamic_field_object_id(
&self,
owner: ObjectID,
name: &str
) -> SuiResult<Option<ObjectID>>

pub fn get_total_transaction_number(&self) -> Result<u64, Error>

pub fn get_transactions_in_range(
&self,
start: TxSequenceNumber,
end: TxSequenceNumber
) -> Result<Vec<(TxSequenceNumber, TransactionDigest)>, Error>

pub fn get_recent_transactions(
&self,
count: u64
) -> Result<Vec<(TxSequenceNumber, TransactionDigest)>, Error>

pub async fn get_transaction(
&self,
digest: TransactionDigest
) -> Result<(VerifiedCertificate, TransactionEffects), Error>

pub fn get_transactions(
&self,
query: TransactionQuery,
cursor: Option<TransactionDigest>,
limit: Option<usize>,
reverse: bool
) -> Result<Vec<TransactionDigest>, Error>

pub fn get_latest_checkpoint_sequence_number(
&self
) -> Result<CheckpointSequenceNumber, Error>

pub fn get_checkpoint_summary_by_sequence_number(
&self,
sequence_number: CheckpointSequenceNumber
) -> Result<CheckpointSummary, Error>

pub fn get_checkpoint_summary_by_digest(
&self,
digest: CheckpointDigest
) -> Result<CheckpointSummary, Error>

pub fn get_checkpoint_contents(
&self,
digest: CheckpointContentsDigest
) -> Result<CheckpointContents, Error>

pub fn get_checkpoint_contents_by_sequence_number(
&self,
sequence_number: CheckpointSequenceNumber
) -> Result<CheckpointContents, Error>

pub async fn get_timestamp_ms(
&self,
digest: &TransactionDigest
) -> Result<Option<u64>, Error>

pub async fn get_events(
&self,
query: EventQuery,
cursor: Option<EventID>,
limit: usize,
descending: bool
) -> Result<Vec<(EventID, SuiEventEnvelope)>, Error>

pub async fn insert_genesis_object(&self, object: Object)

pub fn make_transaction_info(
&self,
transaction_digest: &TransactionDigest,
epoch_store: &Arc<AuthorityPerEpochStore>
) -> Result<Option<VerifiedTransactionInfoResponse>, SuiError>

Make an information response for a transaction

pub fn get_signed_effects_and_maybe_resign(
&self,
cur_epoch: EpochId,
transaction_digest: &TransactionDigest
) -> SuiResult<Option<VerifiedSignedTransactionEffects>>

Get the signed effects of the given transaction. If the effects was signed in a previous epoch, re-sign it so that the caller is able to form a cert of the effects in the current epoch.

pub async fn set_transaction_lock(
&self,
mutable_input_objects: &[ObjectRef],
signed_transaction: VerifiedSignedTransaction,
epoch_store: &Arc<AuthorityPerEpochStore>
) -> Result<(), SuiError>

Set the transaction lock to a specific transaction

pub async fn get_transaction_lock(
&self,
object_ref: &ObjectRef,
epoch_store: &AuthorityPerEpochStore
) -> Result<Option<VerifiedSignedTransaction>, SuiError>

Get the TransactionEnvelope that currently locks the given object, if any. Since object locks are only valid for one epoch, we also need the epoch_id in the query. Returns SuiError::ObjectNotFound if no lock records for the given object can be found. Returns SuiError::ObjectVersionUnavailableForConsumption if the object record is at a different version. Returns Some(VerifiedEnvelope) if the given ObjectRef is locked by a certain transaction. Returns None if the a lock record is initialized for the given ObjectRef but not yet locked by any transaction, or cannot find the transaction in transaction table, because of data race etc.

pub async fn read_certificate(
&self,
digest: &TransactionDigest
) -> Result<Option<VerifiedCertificate>, SuiError>

Read from the DB of certificates

pub async fn parent(&self, object_ref: &ObjectRef) -> Option<TransactionDigest>

pub async fn get_objects(
&self,
_objects: &[ObjectID]
) -> Result<Vec<Option<Object>>, SuiError>

pub async fn get_parent_iterator(
&self,
object_id: ObjectID,
seq: Option<SequenceNumber>
) -> Result<impl Iterator<Item = (ObjectRef, TransactionDigest)> + '_, SuiError>

Returns all parents (object_ref and transaction digests) that match an object_id, at any object version, or optionally at a specific version.

pub async fn get_latest_parent_entry(
&self,
object_id: ObjectID
) -> Result<Option<(ObjectRef, TransactionDigest)>, SuiError>

pub async fn create_advance_epoch_tx_cert(
&self,
epoch_store: &Arc<AuthorityPerEpochStore>,
gas_cost_summary: &GasCostSummary,
epoch_start_timestamp_ms: CheckpointTimestamp,
timeout: Duration,
transaction_certifier: &dyn TransactionCertifier
) -> Result<VerifiedCertificate>