Struct sui_core::authority::AuthorityStore

pub struct AuthorityStore { /* private fields */ }

ALL_OBJ_VER determines whether we want to store all past versions of every object in the store. Authority doesn’t store them, but other entities such as replicas will. S is a template on Authority signature state. This allows SuiDataStore to be used on either authorities or non-authorities. Specifically, when storing transactions and effects, S allows SuiDataStore to either store the authority signed version or unsigned version.

Implementations

impl AuthorityStore

pub async fn open(
path: &Path,
db_options: Option<Options>,
genesis: &Genesis,
committee_store: &Arc<CommitteeStore>,
pruning_config: &AuthorityStorePruningConfig
) -> SuiResult<Self>

Open an authority store by directory path. If the store is empty, initialize it using genesis.

pub async fn open_with_committee_for_testing(
path: &Path,
db_options: Option<Options>,
committee: &Committee,
genesis: &Genesis,
pruning_config: &AuthorityStorePruningConfig
) -> SuiResult<Self>

pub fn epoch_store(&self) -> Guard<Arc<AuthorityPerEpochStore>>

pub fn load_epoch_store(
&self,
intended_epoch: EpochId
) -> SuiResult<Guard<Arc<AuthorityPerEpochStore>>>

pub fn get_effects(
&self,
transaction_digest: &TransactionDigest
) -> SuiResult<TransactionEffects>

Returns the TransactionEffects if we have an effects structure for this transaction digest

pub fn effects_exists(
&self,
transaction_digest: &TransactionDigest
) -> SuiResult<bool>

Returns true if we have an effects structure for this transaction digest

pub fn database_is_empty(&self) -> SuiResult<bool>

Returns true if there are no objects in the database

pub fn get_owner_objects(
&self,
owner: Owner
) -> Result<Vec<ObjectInfo>, SuiError>

pub fn get_owner_objects_iterator(
&self,
owner: Owner
) -> Result<impl Iterator<Item = ObjectInfo> + '_, SuiError>

pub fn get_object_by_key(
&self,
object_id: &ObjectID,
version: VersionNumber
) -> Result<Option<Object>, SuiError>

pub fn object_version_exists(
&self,
object_id: &ObjectID,
version: VersionNumber
) -> Result<bool, SuiError>

pub fn get_object(
&self,
object_id: &ObjectID
) -> Result<Option<Object>, SuiError>

Read an object and return it, or Err(ObjectNotFound) if the object was not found.

pub fn get_objects(
&self,
objects: &[ObjectID]
) -> Result<Vec<Option<Object>>, SuiError>

Get many objects

pub fn check_input_objects(
&self,
objects: &[InputObjectKind]
) -> Result<Vec<Object>, SuiError>

pub async fn get_missing_input_objects(
&self,
digest: &TransactionDigest,
objects: &[InputObjectKind]
) -> Result<Vec<ObjectKey>, SuiError>

When making changes, please see if check_sequenced_input_objects() below needs similar changes as well.

pub fn check_sequenced_input_objects(
&self,
digest: &TransactionDigest,
objects: &[InputObjectKind]
) -> Result<Vec<Object>, SuiError>

When making changes, please see if get_missing_input_objects() above needs similar changes as well.

pub async fn get_object_locking_transaction(
&self,
object_ref: &ObjectRef,
epoch_id: EpochId
) -> SuiResult<Option<VerifiedSignedTransaction>>

Get the TransactionEnvelope that currently locks the given object, if any. Since object locks are only valid for one epoch, we also need the epoch_id in the query. Returns SuiError::ObjectNotFound if no lock records for the given object can be found. Returns SuiError::ObjectVersionUnavailableForConsumption if the object record is at a different version. Returns Some(VerifiedEnvelope) if the given ObjectRef is locked by a certain transaction. Returns None if the a lock record is initialized for the given ObjectRef but not yet locked by any transaction, or cannot find the transaction in transaction table, because of data race etc.

pub fn read_certificate(
&self,
digest: &TransactionDigest
) -> Result<Option<VerifiedCertificate>, SuiError>

Read a certificate and return an option with None if it does not exist.

pub fn parent(
&self,
object_ref: &ObjectRef
) -> Result<Option<TransactionDigest>, SuiError>

Read the transactionDigest that is the parent of an object reference (ie. the transaction that created an object at this version.)

pub fn multi_get_parents(
&self,
object_refs: &[ObjectRef]
) -> Result<Vec<Option<TransactionDigest>>, SuiError>

Batch version of parent function.

pub fn get_parent_iterator(
&self,
object_id: ObjectID,
seq: Option<SequenceNumber>
) -> Result<impl Iterator<Item = (ObjectRef, TransactionDigest)> + '_, SuiError>

Returns all parents (object_ref and transaction digests) that match an object_id, at any object version, or optionally at a specific version.

pub async fn insert_genesis_object(&self, object: Object) -> SuiResult

Insert a genesis object.

pub async fn insert_object_direct(
&self,
object_ref: ObjectRef,
object: &Object
) -> SuiResult

Insert an object directly into the store, and also update relevant tables NOTE: does not handle transaction lock. This is used to insert genesis objects

pub async fn bulk_object_insert(&self, objects: &[&Object]) -> SuiResult<()>

This function is used by the bench.rs script, and should not be used in other contexts In particular it does not check the old locks before inserting new ones, so the objects must be new.

pub async fn lock_and_write_transaction(
&self,
epoch: EpochId,
owned_input_objects: &[ObjectRef],
transaction: VerifiedSignedTransaction
) -> Result<(), SuiError>

Acquires the transaction lock for a specific transaction, writing the transaction to the transaction column family if acquiring the lock succeeds. The lock service is used to atomically acquire locks.

pub async fn update_state(
&self,
inner_temporary_store: InnerTemporaryStore,
certificate: &VerifiedCertificate,
effects: &SignedTransactionEffects,
effects_digest: &TransactionEffectsDigest
) -> SuiResult

Updates the state resulting from the execution of a certificate.

Internally it checks that all locks for active inputs are at the correct version, and then writes objects, certificates, parents and clean up locks atomically.

pub async fn update_objects_state_for_genesis(
&self,
inner_temporary_store: InnerTemporaryStore,
transaction_digest: TransactionDigest
) -> Result<(), SuiError>

Persist temporary storage to DB for genesis modules

pub fn check_locks_exist(&self, objects: &[ObjectRef]) -> SuiResult

Checks multiple object locks exist. Returns SuiError::ObjectNotFound if cannot find lock record for at least one of the objects. Returns SuiError::ObjectVersionUnavailableForConsumption if at least one object lock is not initialized at the given version.

pub async fn revert_state_update(
&self,
tx_digest: &TransactionDigest
) -> SuiResult

This function is called at the end of epoch for each transaction that’s executed locally on the validator but didn’t make to the last checkpoint. The effects of the execution is reverted here. The following things are reverted:

Certificate and effects are deleted.

Latest parent_sync entries for each mutated object are deleted.

All new object states are deleted.

owner_index table change is reverted.

pub fn find_object_lt_or_eq_version(
&self,
object_id: ObjectID,
version: SequenceNumber
) -> Option<Object>

Return the object with version less then or eq to the provided seq number. This is used by indexer to find the correct version of dynamic field child object. We do not store the version of the child object, but because of lamport timestamp, we know the child must have version number less then or eq to the parent.

pub fn get_latest_parent_entry(
&self,
object_id: ObjectID
) -> Result<Option<(ObjectRef, TransactionDigest)>, SuiError>

Returns the last entry we have for this object in the parents_sync index used to facilitate client and authority sync. In turn the latest entry provides the latest object_reference, and also the latest transaction that has interacted with this object.

This parent_sync index also contains entries for deleted objects (with a digest of ObjectDigest::deleted()), and provides the transaction digest of the certificate that deleted the object. Note that a deleted object may re-appear if the deletion was the result of the object being wrapped in another object.

If no entry for the object_id is found, return None.

pub fn object_exists(&self, object_id: ObjectID) -> SuiResult<bool>

pub async fn acquire_shared_locks_from_effects(
&self,
certificate: &VerifiedCertificate,
effects: &TransactionEffects
) -> SuiResult

Lock a sequence number for the shared objects of the input transaction based on the effects of that transaction. Used by full nodes, which don’t listen to consensus.

pub fn get_transaction(
&self,
transaction_digest: &TransactionDigest
) -> SuiResult<Option<VerifiedSignedTransaction>>

pub fn get_certified_transaction(
&self,
transaction_digest: &TransactionDigest
) -> SuiResult<Option<VerifiedCertificate>>

pub fn multi_get_certified_transaction(
&self,
transaction_digests: &[TransactionDigest]
) -> SuiResult<Vec<Option<VerifiedCertificate>>>

pub fn get_sui_system_state_object(&self) -> SuiResult<SuiSystemState>

pub fn transaction_exists(
&self,
cur_epoch: EpochId,
transaction_digest: &TransactionDigest
) -> SuiResult<bool>

Returns true if we have a transaction structure for this transaction digest

pub fn get_signed_transaction_info(
&self,
transaction_digest: &TransactionDigest
) -> Result<VerifiedTransactionInfoResponse, SuiError>