Struct sui_core::authority::AuthorityStore

pub struct AuthorityStore { /* private fields */ }

ALL_OBJ_VER determines whether we want to store all past versions of every object in the store. Authority doesn’t store them, but other entities such as replicas will. S is a template on Authority signature state. This allows SuiDataStore to be used on either authorities or non-authorities. Specifically, when storing transactions and effects, S allows SuiDataStore to either store the authority signed version or unsigned version.

Implementations

impl AuthorityStore

pub async fn open( perpetual_tables: Arc<AuthorityPerpetualTables>, genesis: &Genesis, config: &NodeConfig, registry: &Registry, ) -> SuiResult<Arc<Self>>

Open an authority store by directory path. If the store is empty, initialize it using genesis.

pub fn update_epoch_flags_metrics(&self, old: &[EpochFlag], new: &[EpochFlag])

pub fn clear_object_per_epoch_marker_table( &self, _execution_guard: &RwLockWriteGuard<'_, EpochId>, ) -> SuiResult<()>

pub async fn open_with_committee_for_testing( perpetual_tables: Arc<AuthorityPerpetualTables>, committee: &Committee, genesis: &Genesis, ) -> SuiResult<Arc<Self>>

pub fn open_no_genesis( perpetual_tables: Arc<AuthorityPerpetualTables>, enable_epoch_sui_conservation_check: bool, registry: &Registry, ) -> SuiResult<Arc<Self>>

Open authority store without any operations that require genesis, such as constructing EpochStartConfiguration or inserting genesis objects.

pub fn get_recovery_epoch_at_restart(&self) -> SuiResult<EpochId>

pub fn get_effects( &self, effects_digest: &TransactionEffectsDigest, ) -> SuiResult<Option<TransactionEffects>>

pub fn effects_exists( &self, effects_digest: &TransactionEffectsDigest, ) -> SuiResult<bool>

Returns true if we have an effects structure for this transaction digest

pub fn get_events( &self, event_digest: &TransactionEventsDigest, ) -> Result<Option<TransactionEvents>, TypedStoreError>

pub fn multi_get_events( &self, event_digests: &[TransactionEventsDigest], ) -> SuiResult<Vec<Option<TransactionEvents>>>

pub fn multi_get_effects<'a>( &self, effects_digests: impl Iterator<Item = &'a TransactionEffectsDigest>, ) -> SuiResult<Vec<Option<TransactionEffects>>>

pub fn get_executed_effects( &self, tx_digest: &TransactionDigest, ) -> SuiResult<Option<TransactionEffects>>

pub fn multi_get_executed_effects_digests( &self, digests: &[TransactionDigest], ) -> SuiResult<Vec<Option<TransactionEffectsDigest>>>

Given a list of transaction digests, returns a list of the corresponding effects only if they have been executed. For transactions that have not been executed, None is returned.

pub fn multi_get_executed_effects( &self, digests: &[TransactionDigest], ) -> SuiResult<Vec<Option<TransactionEffects>>>

Given a list of transaction digests, returns a list of the corresponding effects only if they have been executed. For transactions that have not been executed, None is returned.

pub fn is_tx_already_executed( &self, digest: &TransactionDigest, ) -> SuiResult<bool>

pub fn get_marker_value( &self, object_key: FullObjectKey, epoch_id: EpochId, use_object_per_epoch_marker_table_v2: bool, ) -> SuiResult<Option<MarkerValue>>

pub fn get_latest_marker( &self, object_id: FullObjectID, epoch_id: EpochId, use_object_per_epoch_marker_table_v2: bool, ) -> SuiResult<Option<(SequenceNumber, MarkerValue)>>

pub async fn notify_read_root_state_hash( &self, epoch: EpochId, ) -> SuiResult<(CheckpointSequenceNumber, Accumulator)>

Returns future containing the state hash for the given epoch once available

pub fn deprecated_insert_finalized_transactions( &self, digests: &[TransactionDigest], epoch: EpochId, sequence: CheckpointSequenceNumber, ) -> SuiResult

pub fn deprecated_get_transaction_checkpoint( &self, digest: &TransactionDigest, ) -> SuiResult<Option<(EpochId, CheckpointSequenceNumber)>>

pub fn deprecated_multi_get_transaction_checkpoint( &self, digests: &[TransactionDigest], ) -> SuiResult<Vec<Option<(EpochId, CheckpointSequenceNumber)>>>

pub fn database_is_empty(&self) -> SuiResult<bool>

Returns true if there are no objects in the database

pub fn object_exists_by_key( &self, object_id: &ObjectID, version: VersionNumber, ) -> SuiResult<bool>

pub fn multi_object_exists_by_key( &self, object_keys: &[ObjectKey], ) -> SuiResult<Vec<bool>>

pub fn multi_get_objects_by_key( &self, object_keys: &[ObjectKey], ) -> Result<Vec<Option<Object>>, SuiError>

pub fn get_objects( &self, objects: &[ObjectID], ) -> Result<Vec<Option<Object>>, SuiError>

Get many objects

pub fn have_deleted_owned_object_at_version_or_after( &self, object_id: &ObjectID, version: VersionNumber, epoch_id: EpochId, ) -> Result<bool, SuiError>

pub fn bulk_insert_live_objects( perpetual_db: &AuthorityPerpetualTables, live_objects: impl Iterator<Item = LiveObject>, expected_sha3_digest: &[u8; 32], ) -> SuiResult<()>

pub fn set_epoch_start_configuration( &self, epoch_start_configuration: &EpochStartConfiguration, ) -> SuiResult

pub fn get_epoch_start_configuration( &self, ) -> SuiResult<Option<EpochStartConfiguration>>

pub fn write_transaction_outputs( &self, epoch_id: EpochId, tx_outputs: &[Arc<TransactionOutputs>], use_object_per_epoch_marker_table_v2: bool, ) -> SuiResult

Updates the state resulting from the execution of a certificate.

Internally it checks that all locks for active inputs are at the correct version, and then writes objects, certificates, parents and clean up locks atomically.

pub fn acquire_transaction_locks( &self, epoch_store: &AuthorityPerEpochStore, owned_input_objects: &[ObjectRef], tx_digest: TransactionDigest, signed_transaction: Option<VerifiedSignedTransaction>, ) -> SuiResult

pub fn check_owned_objects_are_live(&self, objects: &[ObjectRef]) -> SuiResult

Checks multiple object locks exist. Returns UserInputError::ObjectNotFound if cannot find lock record for at least one of the objects. Returns UserInputError::ObjectVersionUnavailableForConsumption if at least one object lock is not initialized at the given version.

pub fn initialize_live_object_markers( live_object_marker_table: &DBMap<ObjectRef, Option<LockDetailsWrapperDeprecated>>, write_batch: &mut DBBatch, objects: &[ObjectRef], is_force_reset: bool, ) -> SuiResult

pub fn revert_state_update(&self, tx_digest: &TransactionDigest) -> SuiResult

This function is called at the end of epoch for each transaction that’s executed locally on the validator but didn’t make to the last checkpoint. The effects of the execution is reverted here. The following things are reverted:

All new object states are deleted.

owner_index table change is reverted.

NOTE: transaction and effects are intentionally not deleted. It’s possible that if this node is behind, the network will execute the transaction in a later epoch. In that case, we need to keep it saved so that when we receive the checkpoint that includes it from state sync, we are able to execute the checkpoint. TODO: implement GC for transactions that are no longer needed.

pub fn find_object_lt_or_eq_version( &self, object_id: ObjectID, version: SequenceNumber, ) -> SuiResult<Option<Object>>

Return the object with version less then or eq to the provided seq number. This is used by indexer to find the correct version of dynamic field child object. We do not store the version of the child object, but because of lamport timestamp, we know the child must have version number less then or eq to the parent.

pub fn get_latest_object_ref_or_tombstone( &self, object_id: ObjectID, ) -> Result<Option<ObjectRef>, SuiError>

Returns the latest object reference we have for this object_id in the objects table.

The method may also return the reference to a deleted object with a digest of ObjectDigest::deleted() or ObjectDigest::wrapped() and lamport version of a transaction that deleted the object. Note that a deleted object may re-appear if the deletion was the result of the object being wrapped in another object.

If no entry for the object_id is found, return None.

pub fn get_latest_object_ref_if_alive( &self, object_id: ObjectID, ) -> Result<Option<ObjectRef>, SuiError>

Returns the latest object reference if and only if the object is still live (i.e. it does not return tombstones)

pub fn get_latest_object_or_tombstone( &self, object_id: ObjectID, ) -> Result<Option<(ObjectKey, ObjectOrTombstone)>, SuiError>

Returns the latest object we have for this object_id in the objects table.

If no entry for the object_id is found, return None.

pub fn insert_transaction_and_effects( &self, transaction: &VerifiedTransaction, transaction_effects: &TransactionEffects, ) -> Result<(), TypedStoreError>

pub fn multi_insert_transaction_and_effects<'a>( &self, transactions: impl Iterator<Item = &'a VerifiedExecutionData>, ) -> Result<(), TypedStoreError>

pub fn multi_get_transaction_blocks( &self, tx_digests: &[TransactionDigest], ) -> SuiResult<Vec<Option<VerifiedTransaction>>>

pub fn get_transaction_block( &self, tx_digest: &TransactionDigest, ) -> Result<Option<VerifiedTransaction>, TypedStoreError>

pub fn get_sui_system_state_object_unsafe(&self) -> SuiResult<SuiSystemState>

This function reads the DB directly to get the system state object. If reconfiguration is happening at the same time, there is no guarantee whether we would be getting the old or the new system state object. Hence this function should only be called during RPC reads where data race is not a major concern. In general we should avoid this as much as possible. If the intent is for testing, you can use AuthorityState:: get_sui_system_state_object_for_testing.

pub fn expensive_check_sui_conservation<T>( self: &Arc<Self>, type_layout_store: T, old_epoch_store: &AuthorityPerEpochStore, ) -> SuiResult
where T: TypeLayoutStore + Send + Copy,

pub fn maybe_reaccumulate_state_hash( &self, cur_epoch_store: &AuthorityPerEpochStore, new_protocol_version: ProtocolVersion, )

This is a temporary method to be used when we enable simplified_unwrap_then_delete. It re-accumulates state hash for the new epoch if simplified_unwrap_then_delete is enabled.

pub async fn prune_objects_and_compact_for_testing( &self, checkpoint_store: &Arc<CheckpointStore>, rpc_index: Option<&RpcIndexStore>, )