Struct sui_core::authority_aggregator::AuthorityAggregator

pub struct AuthorityAggregator<A> {
    pub committee: Committee,
    pub authority_clients: BTreeMap<AuthorityName, SafeClient<A>>,
    pub metrics: AuthAggMetrics,
    pub safe_client_metrics_base: Arc<SafeClientMetricsBase>,
    pub timeouts: TimeoutConfig,
    pub committee_store: Arc<CommitteeStore>,
}

Fields
committee: Committee

Our Sui committee.

authority_clients: BTreeMap<AuthorityName, SafeClient<A>>

How to talk to this committee.

metrics: AuthAggMetrics

Metrics

safe_client_metrics_base: Arc<SafeClientMetricsBase>

Metric base for the purpose of creating new safe clients during reconfiguration.

timeouts: TimeoutConfigcommittee_store: Arc<CommitteeStore>

Store here for clone during re-config.

Implementations

impl<A> AuthorityAggregator<A>

pub fn new(
committee: Committee,
committee_store: Arc<CommitteeStore>,
authority_clients: BTreeMap<AuthorityName, A>,
registry: &Registry
) -> Self

pub fn new_with_timeouts(
committee: Committee,
committee_store: Arc<CommitteeStore>,
authority_clients: BTreeMap<AuthorityName, A>,
registry: &Registry,
timeouts: TimeoutConfig
) -> Self

pub fn recreate_with_net_addresses(
&self,
committee: CommitteeWithNetAddresses,
network_config: &Config
) -> SuiResult<AuthorityAggregator<NetworkAuthorityClient>>

This function recreates AuthorityAggregator with the given committee. It also updates committee store which impacts other of its references. If it is called on a Validator/Fullnode, it may interleave with the the authority active’s reconfiguration process, and leave the commmittee store in an inconsistent state. When catching up to the latest epoch, it should call reconfig_from_genesis first to fill in all previous epoch’s committee info.

pub fn get_client(&self, name: &AuthorityName) -> Option<&SafeClient<A>>

pub fn clone_client(&self, name: &AuthorityName) -> SafeClient<A>where
A: Clone,

pub fn clone_inner_clients(&self) -> BTreeMap<AuthorityName, A>where
A: Clone,

impl AuthorityAggregator<NetworkAuthorityClient>

pub fn new_from_system_state(
store: &Arc<AuthorityStore>,
committee_store: &Arc<CommitteeStore>,
prometheus_registry: &Registry
) -> Result<Self>

Create a new network authority aggregator by reading the committee and network address information from the system state object on-chain.

impl<A> AuthorityAggregator<A>where
A: AuthorityAPI + Send + Sync + 'static + Clone,

pub async fn get_committee_info(
&self,
epoch: Option<EpochId>
) -> SuiResult<CommitteeInfo>

Query validators for committee information for epoch (None indicates latest epoch) and try to form a CommitteeInfo if we can get a quorum.

pub async fn get_committee_with_net_addresses(
&self,
minimal_epoch: EpochId
) -> SuiResult<CommitteeWithNetAddresses>

Query validators for latest SuiSystemState and try to form CommitteeWithNetworkAddress. Only return Some(CommitteeWithNetAddresses) when there is quorum. This function tolerates uninteresting differences in SuiSystemState as long as they all link to the same CommitteeWithNetAddresses. This function ignores SuiSystemState that has older epoch than minimal_epoch. Usually, the caller should pass in the local incumbent epoch id as minimal_epoch, to avoid getting confused by byzantine validators.

pub async fn process_transaction(
&self,
transaction: VerifiedTransaction
) -> Result<VerifiedCertificate, SuiError>

Submits the transaction to a quorum of validators to make a certificate.

pub async fn process_certificate(
&self,
certificate: CertifiedTransaction
) -> Result<VerifiedCertifiedTransactionEffects, SuiError>

Process a certificate assuming that 2f+1 authorities already are up to date.

This call is meant to be called after process_transaction returns a certificate. At that point (and after) enough authorities are up to date with all objects needed to process the certificate that a submission should succeed. However, in case an authority returns an error, we do try to bring it up to speed.

pub async fn execute_transaction(
&self,
transaction: &VerifiedTransaction
) -> Result<(VerifiedCertificate, VerifiedCertifiedTransactionEffects), Error>

pub async fn get_object_info_execute(
&self,
object_id: ObjectID
) -> SuiResult<ObjectRead>

pub async fn handle_checkpoint_request(
&self,
request: &CheckpointRequest,
authorities: &BTreeSet<AuthorityName>,
timeout_total: Option<Duration>
) -> SuiResult<CheckpointResponse>

pub async fn get_certified_checkpoint(
&self,
sequence_number: CheckpointSequenceNumber,
request_contents: bool,
authorities: &BTreeSet<AuthorityName>,
timeout_total: Option<Duration>
) -> SuiResult<(CertifiedCheckpointSummary, Option<CheckpointContents>)>

pub async fn handle_cert_info_request(
&self,
digest: &TransactionDigest,
timeout_total: Option<Duration>
) -> SuiResult<VerifiedTransactionInfoResponse>

This function tries to fetch CertifiedTransaction from any validators. Returns Error if certificate cannot be found in any validators.

pub async fn handle_transaction_and_effects_info_request(
&self,
digests: &ExecutionDigests,
authorities: Option<&BTreeSet<AuthorityName>>,
timeout_total: Option<Duration>
) -> SuiResult<(VerifiedCertificate, SignedTransactionEffects)>

pub async fn handle_transaction_info_request_from_some_validators(
&self,
tx_digest: &TransactionDigest,
validators: &BTreeSet<AuthorityName>,
timeout_total: Option<Duration>
) -> SuiResult<(Option<VerifiedSignedTransaction>, Option<VerifiedCertificate>)>

This function tries to get SignedTransaction OR CertifiedTransaction from an given list of validators who are supposed to know about it.

pub async fn execute_cert_to_true_effects(
&self,
cert: &CertifiedTransaction
) -> SuiResult<SignedTransactionEffects>

Given a certificate, execute the cert on remote validators (and preferentially on the signers of the cert who are guaranteed to be able to process it immediately) until we receive f+1 identical SignedTransactionEffects - at this point we know we have the true effects for the cert, because of f+1 validators, at least 1 must be honest.

It is assumed that this method will not be called by any of the signers of the cert, since they can simply execute the cert locally and compute their own effects.

pub async fn authorty_ask_for_cert_with_retry_and_timeout(
&self,
transaction: &VerifiedTransaction,
state: &AuthorityState,
timeout: Duration
) -> Result<VerifiedCertificate>