Module sui_core::checkpoints

Modules

checkpoint_executor

CheckpointExecutor is a Node component that executes all checkpoints for the given epoch. It acts as a Consumer to StateSync for newly synced checkpoints, taking these checkpoints and scheduling and monitoring their execution. Its primary goal is to allow for catching up to the current checkpoint sequence number of the network as quickly as possible so that a newly joined, or recovering Node can participate in a timely manner. To that end, CheckpointExecutor attempts to saturate the CPU with executor tasks (one per checkpoint), each of which handle scheduling and awaiting checkpoint transaction execution.

Structs

BuilderCheckpointSummary

CheckpointAggregator

CheckpointBuilder

CheckpointMetrics

CheckpointService

CheckpointServiceNoop

CheckpointSignatureAggregator

CheckpointStore

CheckpointStoreTables

CheckpointStoreTablesConfigurator

Create config structs for configuring DBMap tables

CheckpointStoreTablesReadOnly

Create an intermediate struct used to open the DBMap tables in secondary mode This is only used internally

EpochStats

LogCheckpointOutput

PendingCheckpoint

PendingCheckpointInfo

PendingCheckpointV2Contents

SendCheckpointToStateSync

SubmitCheckpointToConsensus

Enums

CheckpointWatermark

PendingCheckpointV2

Traits

CheckpointServiceNotify

Type Aliases

CheckpointHeight